% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{all_inherit}
\alias{all_inherit}
\title{all_inherit}
\usage{
all_inherit(list_object, class_names)
}
\description{
Test whether all objects within a list all inherit from the same given class.
}
\keyword{internal}

