% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{checkDataFrameClasses}
\alias{checkDataFrameClasses}
\title{checkDataFrameClasses}
\usage{
checkDataFrameClasses(object, matchClasses)
}
\description{
Test whether the classes of each column in a data.frame match the supplied
list of expected class names.
}
\keyword{internal}
