% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Date-generics.R
\name{Cohort}
\alias{Cohort}
\alias{Cohort<-}
\title{Cohort}
\usage{
Cohort(object, value, type)

Cohort(object) <- value
}
\arguments{
\item{object}{The object to get or set the cohorts of, generally a \code{".query"}.}

\item{value}{A \code{"dateRange"} to set the object's cohort to.}

\item{type}{The type of cohort. Only the default of \code{"FIRST_VISIT_DATE"}
is currently valid.}
}
\description{
Get or define a cohort.

Set a cohort of a query.
}
\seealso{
Other date range functions: 
\code{\link{DateRange}()},
\code{\link{EndDate}()},
\code{\link{SplitDateRange}()},
\code{\link{StartDate}()}
}
\concept{date range functions}
