% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/GaGetCoreReport.R
\name{ColTypes}
\alias{ColTypes}
\title{This function applies asFun to the selected columns from the data.frame, df,
matched by colNames (case insensitive).}
\usage{
ColTypes(df, colNames, asFun, ...)
}
\description{
This function applies asFun to the selected columns from the data.frame, df,
matched by colNames (case insensitive).
}

