% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr-classes.R
\docType{class}
\name{.compoundExpr-class}
\alias{.compoundExpr-class}
\title{\code{.compoundExpr} class.}
\description{
An S4 class representing an expression containing one or more conditions.
}
\keyword{internal}
