% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Date-generics.R, R/Date-methods.R
\name{DateRange}
\alias{DateRange}
\alias{DateRange<-}
\alias{DateRange,ANY,ANY-method}
\alias{DateRange,ANY,missing-method}
\alias{DateRange,.standardQuery,missing-method}
\alias{DateRange<-,.standardQuery-method}
\alias{DateRange,gaView,ANY-method}
\alias{DateRange<-,ANY-method}
\title{DateRange}
\usage{
DateRange(object, endDate)

DateRange(object) <- value

\S4method{DateRange}{ANY,ANY}(object, endDate)

\S4method{DateRange}{ANY,missing}(object)

\S4method{DateRange}{.standardQuery,missing}(object)

\S4method{DateRange}{.standardQuery}(object) <- value

\S4method{DateRange}{gaView,ANY}(object)

\S4method{DateRange}{ANY}(object) <- value
}
\arguments{
\item{object}{The start date of the date range or a object to coerce to a
date range. Alternatively, a \code{.query} object to replace the date range
of.}

\item{endDate}{The end date of the date range. Alternatively, if
\code{object} is a \code{.query} object, then \code{endDate} is the
replacement date range.}

\item{value}{The replacement date range.}
}
\description{
Get the date range.

Set the date range.
}
\section{Functions}{
\itemize{
\item \code{DateRange,ANY,ANY-method}: generates a date range object using the supplied
vectors of start date and end dates.

\item \code{DateRange,ANY,missing-method}: Returns the date range of the given query or coerces
the supplied object into a dateRange.

\item \code{DateRange,.standardQuery,missing-method}: Returns the date range of the given query.

\item \code{DateRange<-,.standardQuery-method}: Modify the date range of the given query.

\item \code{DateRange,gaView,ANY-method}: Returns the maximum date range of when a view has been
receiving hits.

\item \code{DateRange<-,ANY-method}: Change the date range of the date range object using
the dates supplied in a vector of length 2, where the first element is the
start date and second being the end date.
}}

\seealso{
Other date range functions: 
\code{\link{Cohort}()},
\code{\link{EndDate}()},
\code{\link{SplitDateRange}()},
\code{\link{StartDate}()}
}
\concept{date range functions}
