% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Date.R, R/GaSegment.R, R/Var.R
\name{GaStartDate}
\alias{GaDateRange}
\alias{GaDateRange<-}
\alias{GaDimensions}
\alias{GaDimensions<-}
\alias{GaEndDate}
\alias{GaEndDate<-}
\alias{GaMetrics}
\alias{GaMetrics<-}
\alias{GaSegment}
\alias{GaSegment<-}
\alias{GaSplitDateRange}
\alias{GaStartDate}
\alias{GaStartDate<-}
\title{Backwards compatibility.}
\usage{
StartDate(object, ...)

EndDate(object, ...)

DateRange(object, endDate)

StartDate(object) <- value

EndDate(object) <- value

DateRange(object) <- value

GaSplitDateRange(dateRange, N)

Segment(object, ..., scope = "sessions")

Segment(object) <- value

Dimensions(object, ...)

Metrics(object, ...)

Dimensions(object) <- value

Metrics(object) <- value
}
\description{
Use \code{Dimensions} instead of \code{GaDimensions}

Use \code{Metrics} instead of \code{GaMetrics}

Use \code{Dimensions<-} instead of \code{GaDimensions<-}

Use \code{Metrics<-} instead of \code{GaMetrics<-}
}

