% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/operand-classes.R
\docType{class}
\name{.dimOperand-class}
\alias{.dimOperand-class}
\title{`dimOperand` class.}
\description{
An S4 class to represent the operand value of a dimension condition expression.
}
\keyword{internal}

