% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Date-generics.R, R/Date-methods.R
\name{EndDate}
\alias{EndDate}
\alias{EndDate<-}
\alias{EndDate,character-method}
\alias{EndDate,dateRange-method}
\alias{EndDate,Interval-method}
\alias{EndDate,.standardQuery-method}
\alias{EndDate,gaView-method}
\alias{EndDate<-,dateRange-method}
\alias{EndDate<-,.standardQuery-method}
\title{EndDate}
\usage{
EndDate(object, value)

EndDate(object) <- value

\S4method{EndDate}{character}(object)

\S4method{EndDate}{dateRange}(object)

\S4method{EndDate}{Interval}(object)

\S4method{EndDate}{.standardQuery}(object)

\S4method{EndDate}{gaView}(object)

\S4method{EndDate}{dateRange}(object) <- value

\S4method{EndDate}{.standardQuery}(object) <- value
}
\arguments{
\item{object}{Object to get end date of.}

\item{value}{Value to set end date of object to.}
}
\description{
Get the end date of the date range.

Set the endDate of the date range.
}
\section{Functions}{
\itemize{
\item \code{EndDate,character-method}: Coerce a character vector into a Google Analytics date
object.

\item \code{EndDate,dateRange-method}: Return the end dates of a date range vector.

\item \code{EndDate,Interval-method}: Return the end dates of a date range vector.

\item \code{EndDate,.standardQuery-method}: Return the end dates of a query's date range vector.

\item \code{EndDate,gaView-method}: Get the last day a view received hits.

\item \code{EndDate<-,dateRange-method}: Set a new end date for a date range.

\item \code{EndDate<-,.standardQuery-method}: Set a new end date for a query.
}}

\seealso{
Other date range functions: 
\code{\link{Cohort}()},
\code{\link{DateRange}()},
\code{\link{SplitDateRange}()},
\code{\link{StartDate}()}
}
\concept{date range functions}
