% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Comparator-methods.R
\docType{methods}
\name{==,.var,.operand-method}
\alias{==,.var,.operand-method}
\title{Equals to}
\usage{
\S4method{==}{.var,.operand}(e1, e2)
}
\arguments{
\item{e1}{Dimension or metric object}

\item{e2}{Operand object}
}
\description{
Do the values on the left and right match exactly.
}
\seealso{
Other comparator_functions: \code{\link{<,.var,.metOperand-method}},
  \code{\link{<=,.var,.metOperand-method}},
  \code{\link{>,.var,.metOperand-method}},
  \code{\link{>=,.var,.metOperand-method}},
  \code{\link{Comparator}}, \code{\link{\%between\%}},
  \code{\link{\%contains\%}}, \code{\link{\%ends_with\%}},
  \code{\link{\%in\%}}, \code{\link{\%matches\%}},
  \code{\link{\%starts_with\%}}, \code{\link{not-equals}}
}
\concept{comparator_functions}
