% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/all-generics.R
\name{Exclude}
\alias{Exclude}
\title{Exclude.}
\usage{
Exclude(object, ...)
}
\arguments{
\item{object}{a condition or sequence to exclude}

\item{...}{further conditions or sequences to exclude}
}
\description{
One or more segment conditions or sequences to exclude from the defined user
or session segment.
}

