% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/management-api-classes.R
\name{GaAccount}
\alias{GaAccount}
\title{GA Account}
\usage{
GaAccount(id = NULL, creds = .creds)
}
\arguments{
\item{id}{ID of the GA account to get}

\item{creds}{The Google APIs credentials to use.}
}
\description{
Get a GA account.
}

