% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/management-api-classes.R
\name{GaAccounts}
\alias{GaAccounts}
\title{GTM Accounts}
\usage{
GaAccounts(creds = .creds)
}
\arguments{
\item{creds}{The Google APIs credentials to use.}
}
\description{
Get the collection of GA accounts accessible to the user with the credentials
supplied by \code{creds}..
}

