% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/management-api-classes.R
\name{GaAccountSummaries}
\alias{GaAccountSummaries}
\title{GTM Account Summaries}
\usage{
GaAccountSummaries(creds = .creds)
}
\arguments{
\item{creds}{The Google APIs credentials to use.}
}
\description{
Get a collection of Google Analytics account summaries.
}

