% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var-classes.R
\docType{class}
\name{gaCalMetVar-class}
\alias{gaCalMetVar-class}
\title{`gaCalMetVar` class.}
\description{
An S4 class to represent a valid Google Analytics Core Reporting API
calculated metric equation.
}
\keyword{internal}
