% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query-classes.R
\docType{class}
\name{gaCohort-class}
\alias{gaCohort-class}
\title{An S4 class to represent a cohort of users over a specific date range}
\description{
An S4 class to represent a cohort of users over a specific date range
}
\section{Slots}{

\describe{
\item{\code{type}}{A length-one character vector representing the type of cohort. At
this time only FIRST_VISIT_DATE is supported by the Google Analytics
Reporting API.}
}}

\keyword{internal}
