% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/comparator-classes.R
\docType{class}
\name{.gaComparator-class}
\alias{.gaComparator-class}
\title{`.gaComparator` class.}
\description{
An S4 class union represent conditional comparator operators.
}
\keyword{internal}

