% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GaSegment.R, R/all-generics.R
\docType{methods}
\name{GaCondition,.compoundExpr-method}
\alias{GaCondition}
\alias{GaCondition,.compoundExpr-method}
\title{GaCondition}
\usage{
\S4method{GaCondition}{.compoundExpr}(object, ..., negation = FALSE)

GaCondition(object, ..., negation = FALSE)
}
\arguments{
\item{object}{An expression to be used as a non-sequential segment condition.}

\item{...}{Other expressions to be ANDed to the first expression provided.}

\item{negation}{Logical TRUE or FALSE to match segments where this conditon
  has not been met.}
}
\description{
Create a new gaSegmentConditionFilter object
}
\section{Methods (by class)}{
\itemize{
\item \code{.compoundExpr}: 
}}

