% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/all-generics.R
\name{GaCreds}
\alias{GaCreds}
\alias{GaCreds<-}
\title{GaCreds}
\usage{
GaCreds(object = "GANALYTICS", creds = NULL, ...)

GaCreds(object) <- value
}
\description{
Get or set the authentication credentials for a Google Analytics query object.

Set the authentication credentials for a Google Analytics query object.
}

