% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/management-api-classes.R
\name{gaCustomDimension-class}
\alias{gaCustomDimension-class}
\title{`gaCustomDimension` class}
\description{
An R6 class representing a defined Google Analytics custom dimension.
}
\keyword{internal}
