% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/all-generics.R
\name{GaDateRange}
\alias{GaDateRange}
\alias{GaDateRange<-}
\title{GaDateRange}
\usage{
GaDateRange(.Object, endDate)

GaDateRange(.Object) <- value
}
\description{
Get the date range.

Set the date range.
}

