% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparator-classes.R
\docType{class}
\name{gaDimComparator-class}
\alias{gaDimComparator-class}
\title{\code{gaDimComparator} class.}
\description{
An S4 class to represent a Core Reporting comparator operator for a dimension
condition.
}
\keyword{internal}
