% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/var-list-classes.R
\docType{class}
\name{gaDimensions-class}
\alias{gaDimensions-class}
\title{`gaDimensions` class.}
\description{
An S4 class to represent a list of Core Reporting dimensions.
}
\keyword{internal}

