% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/deprecated.R
\name{GaDimensions}
\alias{GaDimensions}
\alias{GaDimensions<-}
\title{GaDimensions (Deprecated).}
\usage{
GaDimensions(...)

GaDimensions(...) <- value
}
\arguments{
\item{...}{arguments passed onto \code{Dimensions}}

\item{value}{passed onto \code{Dimensions}}
}
\description{
Use \code{Dimensions} instead of \code{GaDimensions}

Use \code{Dimensions<-} instead of \code{GaDimensions<-}
}

