% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/expression-classes.R
\docType{class}
\name{gaDimExpr-class}
\alias{gaDimExpr-class}
\title{`gaDimExpr` class.}
\description{
An S4 class to represent a Core-Reporting dimension condition expression.
}
\keyword{internal}

