% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-filter-classes.R
\docType{class}
\name{gaDimFilter-class}
\alias{gaDimFilter-class}
\title{`gaDimFilter` class.}
\description{
An S4 class to represent a Core Reporting query table dimension filter expression.
}
\keyword{internal}
