% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/operand-classes.R
\docType{class}
\name{gaDimOperand-class}
\alias{gaDimOperand-class}
\title{`gaDimOperand` class.}
\description{
An S4 class to represent a Core-Reporting dimension operand.
}
\keyword{internal}

