% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/var-classes.R
\docType{class}
\name{gaDimVar-class}
\alias{gaDimVar-class}
\title{`gaDimVar` class.}
\description{
An S4 class to represent a valid Google Analytics Core Reporting API dimension
name.
}
\details{
Set to "ga:date" by default.
}
\keyword{internal}

