% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment-classes.R
\docType{class}
\name{gaDynSegment-class}
\alias{gaDynSegment-class}
\title{`gaDynSegment` class.}
\description{
An S4 class to represent a list of segment filter expressions that belong to a
specified scope.
}
\keyword{internal}
