% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/all-generics.R
\name{GaEndDate}
\alias{GaEndDate}
\alias{GaEndDate<-}
\title{GaEndDate}
\usage{
GaEndDate(.Object)

GaEndDate(.Object) <- value
}
\description{
Get the end date of the date range.

Set the endDate of the date range.
}

