% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/table-filter-classes.R
\docType{class}
\name{gaFilter-class}
\alias{gaFilter-class}
\title{`gaFilter` class.}
\description{
An S4 class to represent a Core Reporting query table filter expression.
}
\keyword{internal}

