% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/deprecated.R
\name{GaFilter}
\alias{GaFilter}
\alias{GaFilter<-}
\title{GaFilter (Deprecated).}
\usage{
GaFilter(...)

GaFilter(...) <- value
}
\arguments{
\item{...}{arguments passed onto \code{TableFilter}}

\item{value}{passed onto \code{TableFilter<-}}
}
\description{
GaFilter (Deprecated).

GaFilter<- (Deprecated).
}
\keyword{internal}

