% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/all-generics.R
\name{GaImmediatelyPrecedes}
\alias{GaImmediatelyPrecedes}
\title{GaImmediatelyPrecedes.}
\usage{
GaImmediatelyPrecedes(object, ...)
}
\arguments{
\item{object}{The expression that should \bold{immediately} preceed others in
the sequence.}

\item{...}{Any other expressions that should \bold{immediately} follow the
  first one but before any others in the sequence.}
}
\description{
Create a gaSegmentSequenceStep object
}
\seealso{
GaSequence
}

