% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/meta.R
\name{GaMetaUpdate}
\alias{GaMetaUpdate}
\title{GaMetaUpdate
Update the metadata file
Update the package system file containing meta data about valid dimensions and metrics, etc.}
\usage{
GaMetaUpdate()
}
\value{
a data.frame
}
\description{
GaMetaUpdate
Update the metadata file
Update the package system file containing meta data about valid dimensions and metrics, etc.
}

