% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/comparator-classes.R
\docType{class}
\name{gaMetComparator-class}
\alias{gaMetComparator-class}
\title{`gaMetComparator` class.}
\description{
An S4 class to represent a comparator operator for a Core Reporting metric
condition.
}
\keyword{internal}

