% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/expression-classes.R
\docType{class}
\name{gaMetExpr-class}
\alias{gaMetExpr-class}
\title{`gaMetExpr` class.}
\description{
An S4 class to represent a Core-Reporting metric condition expression.
}
\keyword{internal}

