\docType{package}
\name{ganalytics}
\alias{ganalytics}
\alias{ganalytics,}
\alias{ganalytics-package}
\title{ganalytics
A Google Analytics client for R
S4 classes and methods for interacting with Google Analytics data in R.}
\description{
  ganalytics A Google Analytics client for R S4 classes and
  methods for interacting with Google Analytics data in R.
}
\references{
  Google Analytics dimensions and metrics reference:
  \url{https://developers.google.com/analytics/devguides/reporting/core/dimsmets}
}
\keyword{package}

