% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/all-generics.R
\name{GaNonSequenceCondition}
\alias{GaNonSequenceCondition}
\title{GaNonSequenceCondition}
\usage{
GaNonSequenceCondition(object, ..., negation = FALSE)
}
\description{
Create a new gaNonSequenceCondition object
}

