% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/deprecated.R
\name{GaNot}
\alias{GaNot}
\title{GaNot (Deprecated, use Not or ! instead).}
\usage{
GaNot(...)
}
\arguments{
\item{...}{arguments passed to \code{Not}}
}
\description{
GaNot (Deprecated, use Not or ! instead).
}

