% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/all-generics.R
\name{GaOperand}
\alias{GaOperand}
\alias{GaOperand<-}
\title{GaOperand}
\usage{
GaOperand(.Object)

GaOperand(.Object) <- value
}
\description{
Get the operand of an expression.

Set the operand of an expression.
}

