% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/all-generics.R
\name{GaOperator}
\alias{GaOperator}
\alias{GaOperator<-}
\title{GaOperator}
\usage{
GaOperator(.Object)

GaOperator(.Object) <- value
}
\description{
Get or create an operator used in an expression.

Set the operator used in an expression.
}

