% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/deprecated.R
\name{GaOr}
\alias{GaOr}
\title{GaOr (Deprecated, use Or or | instead).}
\usage{
GaOr(...)
}
\arguments{
\item{...}{arguments passed to \code{Or}}
}
\description{
GaOr (Deprecated, use Or or | instead).
}
\keyword{internal}

