% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GaSegment.R, R/all-generics.R
\docType{methods}
\name{GaPerHit,gaMetExpr-method}
\alias{GaPerHit}
\alias{GaPerHit,gaMetExpr-method}
\title{GaPerHit}
\usage{
\S4method{GaPerHit}{gaMetExpr}(object, ...)

GaPerHit(object, ...)
}
\arguments{
\item{object}{a gaMetExpr object to coerce to user-level.}

\item{...}{Other conditions to be included in the segment definition.}
}
\description{
Set the scope of a gaMetExpr object to hit-level.
}
\section{Methods (by class)}{
\itemize{
\item \code{gaMetExpr}: 
}}

