% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GaSegment.R, R/all-generics.R
\docType{methods}
\name{GaPerUser,gaSegmentFilterList-method}
\alias{GaPerUser}
\alias{GaPerUser,gaMetExpr-method}
\alias{GaPerUser,gaSegmentFilterList-method}
\title{GaPerUser}
\usage{
\S4method{GaPerUser}{gaSegmentFilterList}(object, ...)

\S4method{GaPerUser}{gaMetExpr}(object, ...)

GaPerUser(object, ...)
}
\arguments{
\item{object}{a gaSegmentFilterList or gaMetExpr object to coerce to user-level.}

\item{...}{Other conditions to be included in the segment definition.}
}
\description{
Set the scope of a gaSegmentFilterList or gaMetExpr object to user-level.
}
\section{Methods (by class)}{
\itemize{
\item \code{gaSegmentFilterList}: 

\item \code{gaMetExpr}: 
}}

