% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query-classes.R
\docType{class}
\name{gaPivot-class}
\alias{gaPivot-class}
\title{An S4 class to represent a pivot report configuration}
\description{
An S4 class to represent a pivot report configuration
}
\section{Slots}{

\describe{
\item{\code{dimensions}}{A list of dimensions to pivot into columns.}

\item{\code{dimensionFilters}}{A filter for one or more of the dimensions.}

\item{\code{metrics}}{A list of metrics to aggregate.}

\item{\code{startGroup}}{The starting index of the group of dimension values to pivot.}

\item{\code{maxGroupCount}}{The maximum number of dimension value groups to pivot over.}
}}

\keyword{internal}
