% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/all-generics.R
\name{GaPrecedes}
\alias{GaPrecedes}
\title{GaPrecedes.}
\usage{
GaPrecedes(object, ...)
}
\arguments{
\item{object}{The expression that should preceed others in the sequence.}

\item{...}{Any other expressions that should follow the first one but before
  any others in the sequence.}
}
\description{
Create a gaSequenceStep object
}
\seealso{
GaSequence
}

