% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/all-generics.R
\name{GaProfileId}
\alias{GaProfileId}
\alias{GaProfileId<-}
\title{GaProfileId}
\usage{
GaProfileId(.Object)

GaProfileId(.Object) <- value
}
\description{
Get the profileId of the query

Set the profileId for the query.
}

