% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query-classes.R
\docType{class}
\name{gaQuery-class}
\alias{gaQuery-class}
\title{\code{gaQuery} class.}
\description{
An S4 class to represent a Core Reporting API query.
}
\section{Slots}{

\describe{
\item{\code{segments}}{A gaSegmentList giving one or more segments to query.}

\item{\code{buckets}}{A numeric vector of breakpoints for histogram buckets.}

\item{\code{lifetimeValue}}{A logical flag of whether to perform life-time value analysis.}
}}

\keyword{internal}
