% Generated by roxygen2 (4.0.2): do not edit by hand
\name{GaQuery}
\alias{GaQuery}
\title{GaQuery
Create a ganalytics query object}
\usage{
GaQuery(profileId, authFile = "~/ganalytics_token.RDS",
  startDate = Sys.Date() - 8, endDate = Sys.Date() - 1,
  metrics = "ga:sessions", dimensions = "ga:date", sortBy = NULL,
  filters = NULL, segment = NULL, samplingLevel = "DEFAULT",
  maxResults = kGaMaxResults)
}
\arguments{
\item{profileId}{profile id to use}

\item{authFile}{auth file to save token to}

\item{startDate}{start date}

\item{endDate}{end date}

\item{metrics}{character vector of metrics}

\item{dimensions}{character vector of dimensions}

\item{sortBy}{a sort by object}

\item{filters}{a filters object}

\item{segment}{a segment object}

\item{samplingLevel}{either "DEFAULT", "HIGHER_PRECISION" or "FASTER"}

\item{maxResults}{the maximum number of results to return,
up to 1,000,000}
}
\description{
GaQuery
Create a ganalytics query object
}

