% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/all-generics.R
\name{GaSamplingLevel}
\alias{GaSamplingLevel}
\alias{GaSamplingLevel<-}
\title{GaSamplingLevel}
\usage{
GaSamplingLevel(.Object)

GaSamplingLevel(.Object) <- value
}
\description{
Get the sampling level.

Set the sampling level for a ganalytics query.
}

