% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Expr.R, R/GaSegment.R, R/all-generics.R
\docType{methods}
\name{GaScopeLevel,gaSegMetExpr-method}
\alias{GaScopeLevel}
\alias{GaScopeLevel,gaSegMetExpr-method}
\alias{GaScopeLevel,gaSegmentCondition-method}
\alias{GaScopeLevel<-}
\alias{GaScopeLevel<-,gaMetExpr,character-method}
\alias{GaScopeLevel<-,gaSegMetExpr,character-method}
\alias{GaScopeLevel<-,gaSegmentCondition,character-method}
\title{GaScopeLevel.}
\usage{
\S4method{GaScopeLevel}{gaSegMetExpr}(object)

\S4method{GaScopeLevel}{gaSegMetExpr,character}(object) <- value

\S4method{GaScopeLevel}{gaMetExpr,character}(object) <- value

\S4method{GaScopeLevel}{gaSegmentCondition}(object)

\S4method{GaScopeLevel}{gaSegmentCondition,character}(object) <- value

\S4method{GaScopeLevel}{gaSegmentCondition}(object)

\S4method{GaScopeLevel}{gaSegmentCondition,character}(object) <- value

GaScopeLevel(object, value)

GaScopeLevel(object) <- value
}
\arguments{
\item{object}{Segment condition or combined segment conditions or metric
expression.}

\item{value}{If a new scope level is supplied, then this function will return
  an updated copy of the supplied object with the new scope applied.}
}
\description{
Get the scope level of a gaDynSegment or gaMetExpr

Set the scope level of a gaDynSegment or a gaMetExpr
For segments, one of 'users' or 'sessions'
For metric expressions one of 'perUser', 'perSession' or 'perHit'
}
\section{Methods (by class)}{
\itemize{
\item \code{gaSegMetExpr}: 

\item \code{object = gaSegMetExpr,value = character}: 

\item \code{object = gaMetExpr,value = character}: 

\item \code{gaSegmentCondition}: 

\item \code{object = gaSegmentCondition,value = character}: 

\item \code{gaSegmentCondition}: 

\item \code{object = gaSegmentCondition,value = character}: 
}}

