% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/segment-classes.R
\docType{class}
\name{.gaSegment-class}
\alias{.gaSegment-class}
\title{`.gaSegment` class.}
\description{
An S4 class union representing dynamic and pre-defined segment expressions.
}
\keyword{internal}

