% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/segment-classes.R
\docType{class}
\name{gaSegmentCondition-class}
\alias{gaSegmentCondition-class}
\title{`gaSegmentCondition` class.}
\description{
An S4 class to represent an expression for a custom segment.
}
\keyword{internal}

