% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/segment-classes.R
\docType{class}
\name{.gaSegmentFilter-class}
\alias{.gaSegmentFilter-class}
\title{`.gaSegmentFilter` class.}
\description{
An S4 superclass to represent a negatable filter expression for a custom
segment.
}
\keyword{internal}

