% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/segment-classes.R
\docType{class}
\name{gaSegmentFilterList-class}
\alias{gaSegmentFilterList-class}
\title{`gaSegmentFilterList` class.}
\description{
An S4 class to represent a list of segment filter expressions that belong to a
specified scope.
}
\keyword{internal}

