% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GaSegment.R, R/all-generics.R
\docType{methods}
\name{GaSegmentFilters,.compoundExpr-method}
\alias{GaSegmentFilters}
\alias{GaSegmentFilters,.compoundExpr-method}
\alias{GaSegmentFilters,.gaSegmentFilter-method}
\alias{GaSegmentFilters,gaSegmentFilterList-method}
\title{GaSegmentFilters}
\usage{
\S4method{GaSegmentFilters}{.compoundExpr}(object, ..., scope = "sessions")

\S4method{GaSegmentFilters}{.gaSegmentFilter}(object, ..., scope = "sessions")

\S4method{GaSegmentFilters}{gaSegmentFilterList}(object)

GaSegmentFilters(object, ..., scope = "sessions")
}
\arguments{
\item{object}{The first condition to be included in the segments definition.}

\item{...}{Other conditions to be included in the segment definition.}

\item{scope}{The scope of this condition, either 'user' or 'session' level.}
}
\description{
Create a new gaSegmentFilterList object
}
\details{
A segment condition is either sequential or non-sequential. Sequential and
non-sequential conditoins can be combined using this function.
}
\section{Methods (by class)}{
\itemize{
\item \code{.compoundExpr}: 

\item \code{.gaSegmentFilter}: 

\item \code{gaSegmentFilterList}: 
}}

