% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/segment-classes.R
\docType{class}
\name{gaSegmentId-class}
\alias{gaSegmentId-class}
\title{`gaSegmentId` class.}
\description{
An S4 class to represent a pre-defined custom or built-in segment by its ID.
}
\keyword{internal}

