% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/segment-classes.R
\docType{class}
\name{gaSegmentList-class}
\alias{gaSegmentList-class}
\title{`gaSegmentList` class.}
\description{
An S4 class to represent a list of segment expressions to query.
}
\keyword{internal}

