% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/segment-classes.R
\docType{class}
\name{gaSegmentSequenceStep-class}
\alias{gaSegmentSequenceStep-class}
\title{`gaSegmentSequenceStep` class.}
\description{
An S4 class to represent a step within a sequence expression.
}
\keyword{internal}

