% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment-classes.R
\docType{class}
\name{gaSegmentsList-class}
\alias{gaSegmentsList-class}
\title{`gaSegmentsList` class.}
\description{
An S4 class to represent a list of segment expressions to query.
}
\keyword{internal}
