% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr-classes.R
\docType{class}
\name{gaSegMetExpr-class}
\alias{gaSegMetExpr-class}
\title{`gaSegMetExpr` class.}
\description{
An S4 class to represent a Core-Reporting metric condition for use in segment
expressions.
}
\keyword{internal}

