% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/deprecated.R
\name{GaSequence}
\alias{GaSequence}
\title{Use \code{Sequence} instead of \code{GaSequence}}
\usage{
GaSequence(...)
}
\arguments{
\item{...}{passed onto \code{Sequence}}
}
\description{
Use \code{Sequence} instead of \code{GaSequence}
}

