% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GaSegment.R, R/all-generics.R
\docType{methods}
\name{GaPrecedes,.compoundExpr-method}
\alias{GaImmediatelyPrecedes}
\alias{GaImmediatelyPrecedes,.compoundExpr-method}
\alias{GaPrecedes}
\alias{GaPrecedes,.compoundExpr-method}
\alias{GaSequenceCondition}
\alias{GaSequenceCondition,.compoundExpr-method}
\alias{GaSequenceCondition,gaSequenceStep-method}
\alias{GaStartsWith}
\alias{GaStartsWith,.compoundExpr-method}
\title{GaPrecedes.}
\usage{
\S4method{GaPrecedes}{.compoundExpr}(object)

\S4method{GaImmediatelyPrecedes}{.compoundExpr}(object)

\S4method{GaStartsWith}{.compoundExpr}(object)

\S4method{GaSequenceCondition}{.compoundExpr}(object, ..., negation = FALSE)

\S4method{GaSequenceCondition}{gaSequenceStep}(object, ..., negation = FALSE)

GaPrecedes(object, ...)

GaImmediatelyPrecedes(object, ...)

GaStartsWith(object, ...)

GaSequenceCondition(object, ..., negation = FALSE)
}
\description{
Create a gaSequenceStep object

Create a gaSequenceStep object

Create a gaSequenceStep object

Create a new gaSequenceCondition object
}
\section{Methods (by class)}{
\itemize{
\item \code{.compoundExpr}: 

\item \code{.compoundExpr}: 

\item \code{.compoundExpr}: 

\item \code{.compoundExpr}: 

\item \code{gaSequenceStep}: 
}}

