% Generated by roxygen2 (4.0.2): do not edit by hand
\name{GaSortBy}
\alias{GaSortBy}
\alias{GaSortBy<-}
\title{GaSortBy}
\usage{
GaSortBy(.Object, ..., desc = logical(0))

GaSortBy(.Object) <- value
}
\description{
Get the sortBy order of the query.

Set the order of rows returned by Google Analytics.
}

