% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Date.R
\name{GaSplitDateRange}
\alias{GaSplitDateRange}
\title{GaSplitDateRange (Deprecated)}
\usage{
GaSplitDateRange(...)
}
\arguments{
\item{...}{arguments passed onto \code{SplitDateRange}}
}
\description{
GaSplitDateRange (Deprecated)
}

