% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/all-generics.R
\name{GaStartDate}
\alias{GaStartDate}
\alias{GaStartDate<-}
\title{GaStartDate}
\usage{
GaStartDate(.Object)

GaStartDate(.Object) <- value
}
\description{
Get the start date.

Set the start date.
}

