% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/management-api-classes.R
\name{gaUserSegment-class}
\alias{gaUserSegment-class}
\title{`gaUserSegment` class}
\description{
An R6 class representing a default or custom segment available with the
credentials provided.
}
\keyword{internal}

