% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/management-api-classes.R
\name{GaUserSegment}
\alias{GaUserSegment}
\title{User Segment}
\usage{
GaUserSegment(id = NULL, definition = NA, creds = .creds)
}
\arguments{
\item{id}{ID of the user or system segment to get}

\item{definition}{The definition to use if an ID is not provided.}

\item{creds}{The Google APIs credentials to use.}
}
\description{
Get a user or system defined segment.
}

