% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/management-api-classes.R
\name{GaUserSegments}
\alias{GaUserSegments}
\title{Get a collection of user and system defined segments.}
\usage{
GaUserSegments(creds = .creds)
}
\arguments{
\item{creds}{The Google APIs credentials to use.}
}
\description{
Returns a collection object of user and system defined segments available
under the supplied user credentials.
}

