% Generated by roxygen2 (4.0.1): do not edit by hand
\name{GaVar}
\alias{GaVar}
\alias{GaVar<-}
\title{GaVar}
\usage{
GaVar(.Object)

GaVar(.Object) <- value
}
\description{
Gets or Creates an object from the superclass .gaVar

Sets the value of an object or sets its slot belonging to the superclass .gaVar
}

