% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Query-generics.R, R/GaView-methods.R
\docType{methods}
\name{GaView}
\alias{GaView}
\alias{GaView,.query-method}
\alias{GaView,character-method}
\alias{GaView,gaAccount-method}
\alias{GaView,gaProperty-method}
\alias{GaView,gaView-method}
\alias{GaView,numeric-method}
\alias{GaView,viewId-method}
\alias{GaView<-}
\alias{GaView<-,.query-method}
\title{GaView.}
\usage{
GaView(object, value)

GaView(object) <- value

\S4method{GaView}{viewId}(object)

\S4method{GaView}{gaView}(object)

\S4method{GaView}{gaProperty}(object)

\S4method{GaView}{gaAccount}(object)

\S4method{GaView}{character}(object)

\S4method{GaView}{numeric}(object)

\S4method{GaView}{.query}(object)

\S4method{GaView}{.query}(object) <- value
}
\arguments{
\item{object}{An object to coerce to a gaView class object or to get the
gaView of, such as a query, default view of a web property, or the default
view of the first web property in a Google Analytics account.}

\item{value}{The optional replacement view if the object supplied is a query,
in which case GaView will return the modified query.}
}
\description{
Get the viewId of the query

Set the viewId for the query.
}
\section{Methods (by class)}{
\itemize{
\item \code{viewId}: Returns itself

\item \code{gaView}: Returns the ID of the supplied view.

\item \code{gaProperty}: Returns the ID of the first view within the supplied
property.

\item \code{gaAccount}: Returns the ID of the first view within the first property
of the supplied account.

\item \code{character}: Parses the supplied character value as a valid view ID.

\item \code{numeric}: Parses the supplied numeric value as a valid view ID.

\item \code{.query}: Returns the view ID of the given query.

\item \code{.query}: Replace the view being used by a query.
}}

