% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GetGa.R
\docType{methods}
\name{GetGaData,.query-method}
\alias{GetGaData,.query-method}
\title{GetGaData
Execute a ganalytics query.}
\usage{
\S4method{GetGaData}{.query}(object, creds = NULL, .progress = "time",
  use_oob = FALSE, addViewId = FALSE)
}
\arguments{
\item{.progress}{progress bar to display. use .progress = "none" to turn off.}

\item{query}{the query to execute.}
}
\value{
a dataframe
}
\description{
GetGaData
Execute a ganalytics query.
}

