% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/GetGa.R
\docType{methods}
\name{GetGaData,gaQuery-method}
\alias{GetGaData,gaQuery-method}
\title{GetGaData
Execute a ganalytics query.}
\usage{
\S4method{GetGaData}{gaQuery}(object, creds = NULL,
  scope = "https://www.googleapis.com/auth/analytics.readonly",
  quiet = FALSE, details = FALSE, .progress = "time", use_oob = FALSE,
  addViewId = FALSE)
}
\arguments{
\item{quiet}{supress messages}

\item{details}{show detailed messages}

\item{.progress}{progress bar to display. use .progress = "none" to turn off.}

\item{query}{the query to execute.}
}
\value{
a dataframe
}
\description{
GetGaData
Execute a ganalytics query.
}

