% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/all-generics.R
\name{GetGaQueries}
\alias{GetGaQueries}
\title{GetGaQueries.}
\usage{
GetGaQueries(object)
}
\arguments{
\item{object}{a query object to coerce to matrix.}
}
\description{
Get the chracter string query compoents for the given ganalytics object.
}

