% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/all-generics.R
\name{GetGaUrl}
\alias{GetGaQueries}
\alias{GetGaUrl}
\title{GetGaUrl}
\usage{
GetGaUrl(.Object)

GetGaQueries(.Object)
}
\description{
Get the utf8 URL string compoent for the given ganalytics object.

Get the chracter string query compoents for the given ganalytics object.
}

