% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GaApiRequest.R
\name{GoogleApiCreds}
\alias{GoogleApiCreds}
\title{Google APIs OAuth 2.0 Credentials.}
\usage{
GoogleApiCreds(userName = character(0), appCreds = NULL,
  cache = character(0), use_oob = FALSE, appname = "GOOGLE_APIS")
}
\arguments{
\item{userName}{Google username email address hint}

\item{appCreds}{Filename or named vector for client_id and client_secret.}

\item{cache}{httr OAuth2.0 cache}

\item{use_oob}{as per httr}

\item{appname}{prefix of environment variables that hold the client ID and client secret.}
}
\description{
Create a Google APIs OAuth2.0 credentials object
}

