% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Comparator-generics.R, R/Comparator-methods.R
\docType{methods}
\name{\%between\%}
\alias{\%between\%}
\alias{\%between\%,.var,.operand-method}
\title{\%between\%}
\usage{
var \%between\% operand

\S4method{\%between\%}{.var,.operand}(var, operand)
}
\arguments{
\item{var}{A metric object or an interval dimension object.}

\item{operand}{A vector of length-two given the lower and upper bounds of the
range.}
}
\description{
A condition where the var (LHS) is within the lower and upper bounds
specified by first and second vector value (respectively) of the operand
(RHS).
}
\seealso{
Other comparator_functions: \code{\link{<,.var,.metOperand-method}},
  \code{\link{<=,.var,.metOperand-method}},
  \code{\link{==,.var,.operand-method}},
  \code{\link{>,.var,.metOperand-method}},
  \code{\link{>=,.var,.metOperand-method}},
  \code{\link{Comparator}}, \code{\link{\%contains\%}},
  \code{\link{\%ends_with\%}}, \code{\link{\%in\%}},
  \code{\link{\%matches\%}}, \code{\link{\%starts_with\%}},
  \code{\link{not-equals}}
}
\concept{comparator_functions}
