% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Comparator-generics.R, R/Comparator-methods.R
\docType{methods}
\name{\%ends_with\%}
\alias{\%ends_with\%}
\alias{\%ends_with\%,.var,.dimOperand-method}
\title{\%ends_with\%}
\usage{
var \%ends_with\% operand

\S4method{\%ends_with\%}{.var,.dimOperand}(var, operand)
}
\arguments{
\item{var}{Dimension object}

\item{operand}{Operand dimension object or a length-one character string}
}
\description{
A condition where the dimension (LHS) matches values that end with
the character string given by the operand (RHS).
}
\seealso{
Other comparator_functions: \code{\link{<,.var,.metOperand-method}},
  \code{\link{<=,.var,.metOperand-method}},
  \code{\link{==,.var,.operand-method}},
  \code{\link{>,.var,.metOperand-method}},
  \code{\link{>=,.var,.metOperand-method}},
  \code{\link{Comparator}}, \code{\link{\%between\%}},
  \code{\link{\%contains\%}}, \code{\link{\%in\%}},
  \code{\link{\%matches\%}}, \code{\link{\%starts_with\%}},
  \code{\link{not-equals}}
}
\concept{comparator_functions}
