% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Comparator-generics.R, R/Comparator-methods.R
\docType{methods}
\name{\%in\%}
\alias{\%in\%}
\alias{\%in\%,.var,.operand-method}
\title{\%in\%}
\usage{
x \%in\% table

\S4method{\%in\%}{.var,.operand}(x, table)
}
\arguments{
\item{x}{Dimension or metric object}

\item{table}{Operand object}
}
\description{
A condition where the dimension (LHS) matches one of the
values in the vector specified by the operand (RHS).
}
\seealso{
Other comparator_functions: \code{\link{<,.var,.metOperand-method}},
  \code{\link{<=,.var,.metOperand-method}},
  \code{\link{==,.var,.operand-method}},
  \code{\link{>,.var,.metOperand-method}},
  \code{\link{>=,.var,.metOperand-method}},
  \code{\link{Comparator}}, \code{\link{\%between\%}},
  \code{\link{\%contains\%}}, \code{\link{\%ends_with\%}},
  \code{\link{\%matches\%}}, \code{\link{\%starts_with\%}},
  \code{\link{not-equals}}
}
\concept{comparator_functions}
