% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Comparator-generics.R, R/Comparator-methods.R
\docType{methods}
\name{\%matches\%}
\alias{\%matches\%}
\alias{\%matches\%,.var,.dimOperand-method}
\title{\%matches\%}
\usage{
var \%matches\% operand

\S4method{\%matches\%}{.var,.dimOperand}(var, operand)
}
\arguments{
\item{var}{A dimension object.}

\item{operand}{A regular expression string.}
}
\description{
A condition where the dimension (LHS) matches a regular
expression given by the operand (RHS).
}
\seealso{
Other comparator_functions: \code{\link{<,.var,.metOperand-method}},
  \code{\link{<=,.var,.metOperand-method}},
  \code{\link{==,.var,.operand-method}},
  \code{\link{>,.var,.metOperand-method}},
  \code{\link{>=,.var,.metOperand-method}},
  \code{\link{Comparator}}, \code{\link{\%between\%}},
  \code{\link{\%contains\%}}, \code{\link{\%ends_with\%}},
  \code{\link{\%in\%}}, \code{\link{\%starts_with\%}},
  \code{\link{not-equals}}
}
\concept{comparator_functions}
