% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Comparator-methods.R
\docType{methods}
\name{>,.var,.metOperand-method}
\alias{>,.var,.metOperand-method}
\title{Greater than}
\usage{
\S4method{>}{.var,.metOperand}(e1, e2)
}
\arguments{
\item{e1}{Dimension or metric object}

\item{e2}{Operand object}
}
\description{
Is the value on the left greater than the value on the right.
}
\seealso{
Other comparator_functions: \code{\link{<,.var,.metOperand-method}},
  \code{\link{<=,.var,.metOperand-method}},
  \code{\link{==,.var,.operand-method}},
  \code{\link{>=,.var,.metOperand-method}},
  \code{\link{Comparator}}, \code{\link{\%between\%}},
  \code{\link{\%contains\%}}, \code{\link{\%ends_with\%}},
  \code{\link{\%in\%}}, \code{\link{\%matches\%}},
  \code{\link{\%starts_with\%}}, \code{\link{not-equals}}
}
\concept{comparator_functions}
