% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/all-generics.R
\name{IsNegated}
\alias{IsNegated}
\title{IsNegated}
\usage{
IsNegated(object)
}
\arguments{
\item{object}{an object to test for negation belonging to the superclass
  \code{.gaSegmentFilter}.}
}
\description{
Tests whether a segment filter is negated.
}

