% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Operator.R, R/all-generics.R
\docType{methods}
\name{IsRegEx,.dimOperator-method}
\alias{IsRegEx}
\alias{IsRegEx,.dimOperator-method}
\alias{IsRegEx,.expr-method}
\title{IsRegEx.}
\usage{
\S4method{IsRegEx}{.dimOperator}(object)

\S4method{IsRegEx}{.expr}(object)

IsRegEx(object)
}
\arguments{
\item{object}{An object to check if whether a regular expression.}
}
\value{
TRUE or FALSE
}
\description{
Checks for a regular expression.
}
\section{Methods (by class)}{
\itemize{
\item \code{.dimOperator}: 

\item \code{.expr}: 
}}

