% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/helper-functions.R
\name{IsVarMatch}
\alias{IsVarMatch}
\title{IsVarMatch
The following method is a temporary workaround to support XX placeholders in dimension and metric
names, such as with custom dimensions, metrics and various goal related variables.}
\usage{
IsVarMatch(thisVar, inVars)
}
\description{
IsVarMatch
The following method is a temporary workaround to support XX placeholders in dimension and metric
names, such as with custom dimensions, metrics and various goal related variables.
}

