% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sequence-generics.R
\name{Later}
\alias{Later}
\title{Later}
\usage{
Later(object, ...)
}
\arguments{
\item{object}{The expression that should precede others in the sequence.}

\item{...}{Any other expressions that should follow the first one but before
any others in the sequence.}
}
\value{
A \code{gaSegmentSequenceStep} object, with the immediate flag not set.
}
\description{
Treat a step within a sequence as happening at any point after any preceding
steps in the sequence, i.e. 'later'. 'Later' means 'followed by', but not
necessarily immediately.
}
\examples{
purchased_sometime_later <- Later(Expr(~transactions > 0))

}
\seealso{
Other sequence segment functions: \code{\link{First}},
  \code{\link{Sequence}}, \code{\link{Then}}
}
\concept{sequence segment functions}
