% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/operand-classes.R
\docType{class}
\name{.logicalDimOperand-class}
\alias{.logicalDimOperand-class}
\alias{.regexDimOperand-class}
\title{`.logicalDimOperand` class.}
\description{
An S4 class to represent the operand value of a logical dimension condition
expression.

An S4 class to represent the regex operand value of a dimension condition
expression.
}
\keyword{internal}

