% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Query.R, R/all-generics.R, R/deprecated.R
\docType{methods}
\name{MaxResults,.query-method}
\alias{GaMaxResults}
\alias{GaMaxResults<-}
\alias{MaxResults}
\alias{MaxResults,.query-method}
\alias{MaxResults<-}
\alias{MaxResults<-,.query-method}
\title{MaxResults.}
\usage{
\S4method{MaxResults}{.query}(object)

\S4method{MaxResults}{.query}(object) <- value

MaxResults(object, value)

MaxResults(object) <- value

GaMaxResults(...)

GaMaxResults(...) <- value
}
\arguments{
\item{object}{a query object.}

\item{value}{replacement value for the max-results parameter of the query.}

\item{...}{arguments passed onto \code{MaxResults}}

\item{value}{passed onto \code{MaxResults}}
}
\description{
Get the value set for MaxResults.

Set the maximum rows returned by a ganalytics query.

GaMaxResults (Deprecated).

GaMaxResults<- (Deprecated).
}
\section{Methods (by class)}{
\itemize{
\item \code{.query}: 

\item \code{.query}: 
}}

