% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparator-classes.R
\docType{class}
\name{.mcfComparator-class}
\alias{.mcfComparator-class}
\title{\code{.mcfComparator} class.}
\description{
An S4 class union representing Multi-Channel Funnel condition comparator
operators.
}
\keyword{internal}
