% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/comparator-classes.R
\docType{class}
\name{mcfDimComparator-class}
\alias{mcfDimComparator-class}
\title{`mcfDimComparator` class.}
\description{
An S4 class representing Multi-Channel Funnel dimension conditional comparator
operators.
}
\keyword{internal}

