% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var-list-classes.R
\docType{class}
\name{mcfDimensions-class}
\alias{mcfDimensions-class}
\title{\code{mcfDimensions} class.}
\description{
An S4 class to represent a list of Multi-Channel Funnel dimensions.
}
\keyword{internal}
