% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/operand-classes.R
\docType{class}
\name{mcfDimOperand-class}
\alias{mcfDimOperand-class}
\title{`mcfDimOperand` class.}
\description{
An S4 class to represent a Multi-Channel dimension condition operand.
}
\keyword{internal}

