% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/var-classes.R
\docType{class}
\name{mcfDimVar-class}
\alias{mcfDimVar-class}
\title{`mcfDimVar` class.}
\description{
An S4 class to represent a Multi-Channel Funnel dimension.
}
\details{
Set to "mcf:nthDay" by default.
}
\keyword{internal}

