% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Expr.R, R/all-generics.R
\docType{methods}
\name{McfExpr,character,character-method}
\alias{McfExpr}
\alias{McfExpr,character,character-method}
\title{McfExpr.}
\usage{
\S4method{McfExpr}{character,character}(object, operator, operand)

McfExpr(object, operator, operand)
}
\arguments{
\item{object}{A dimension or metric variable, or another object to be coerced
to an .expr object.}

\item{operator}{The operator to use for the expression.}

\item{operand}{The operand to use for the expression.}
}
\description{
Create a Multi-Chanel Funnel Reporting API expression.
}
\section{Methods (by class)}{
\itemize{
\item \code{object = character,operator = character}: 
}}
\examples{
\donttest{
  myQuery <- McfQuery(view = 123456789)
  source_matches_google <- McfExpr("mcf:source", "~", "google")
  TableFilter(myQuery) <- source_matches_google
  GetData(myQuery)
}
}

