% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/table-filter-classes.R
\docType{class}
\name{mcfFilter-class}
\alias{mcfFilter-class}
\title{`mcfFilter` class.}
\description{
An S4 class to represent Multi-Channel Funnel query table filter expression.
}
\keyword{internal}

