% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/comparator-classes.R
\docType{class}
\name{mcfMetComparator-class}
\alias{mcfMetComparator-class}
\title{`mcfMetComparator` class.}
\description{
An S4 class representing Multi-Channel Funnel metric condtion comparison
operators.
}
\keyword{internal}

