% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/expression-classes.R
\docType{class}
\name{mcfMetExpr-class}
\alias{mcfMetExpr-class}
\title{`mcfMetExpr` class.}
\description{
An S4 class to represent a Multi-Channel Funnel metric condition expression.
}
\keyword{internal}

