% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/all-generics.R
\name{McfMetrics}
\alias{McfMetrics}
\alias{McfMetrics<-}
\title{McfMetrics}
\usage{
McfMetrics(.Object, ...)

McfMetrics(.Object) <- value
}
\description{
Get the metrics of the object.

Set the metrics of the object.
}

