% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/var-classes.R
\docType{class}
\name{mcfMetVar-class}
\alias{mcfMetVar-class}
\title{`mcfMetVar` class.}
\description{
An S4 class to represent a Multi-Channel Funnel metric.
}
\details{
Set to "mcf:totalConversions" by default.
}
\keyword{internal}

