% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operand-classes.R
\docType{class}
\name{.mcfOperand-class}
\alias{.mcfOperand-class}
\title{\code{.mcfOperand} class.}
\description{
An S4 class union representing Multi-Channel Funnel condition operands.
}
\keyword{internal}
