% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/all-generics.R
\name{McfOperand}
\alias{McfOperand}
\alias{McfOperand<-}
\title{McfOperand}
\usage{
McfOperand(.Object)

McfOperand(.Object) <- value
}
\description{
Get the operand of an expression.

Set the operand of an expression.
}

