% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Query.R
\name{McfQuery}
\alias{McfQuery}
\title{McfQuery.}
\usage{
McfQuery(view = NA, creds = .creds, startDate = Sys.Date() - 8,
  endDate = Sys.Date() - 1, metrics = "mcf:totalConversions",
  dimensions = "mcf:nthDay", sortBy = NULL, filters = NULL,
  samplingLevel = "DEFAULT", maxResults = kGaMaxResults)
}
\arguments{
\item{view}{view id to use}

\item{creds}{authentication credentials object created using GoogleApiCreds()}

\item{startDate}{start date}

\item{endDate}{end date}

\item{metrics}{character vector of metrics}

\item{dimensions}{character vector of dimensions}

\item{sortBy}{a sort by object}

\item{filters}{a filters object}

\item{samplingLevel}{either "DEFAULT", "HIGHER_PRECISION" or "FASTER"}

\item{maxResults}{the maximum number of results to return,
 up to 1,000,000}
}
\description{
Create a Multi-Channel Funnel Reporting API query object
}

