% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/all-generics.R
\name{McfSortBy}
\alias{McfSortBy}
\alias{McfSortBy<-}
\title{McfSortBy}
\usage{
McfSortBy(.Object, ..., desc = logical(0))

McfSortBy(.Object) <- value
}
\description{
Get the sortBy order of the query.

Set the order of rows returned by Google Analytics.
}

