% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/all-generics.R
\name{Var<-}
\alias{McfVar}
\alias{McfVar<-}
\alias{Var<-}
\title{Var<-}
\usage{
Var(.Object) <- value

McfVar(.Object)

McfVar(.Object) <- value
}
\description{
Sets the value of an object or sets its slot belonging to the superclass .mcfVar

Gets or Creates an object from the superclass .mcfVar

Sets the value of an object or sets its slot belonging to the superclass .mcfVar
}

