% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparator-classes.R
\docType{class}
\name{.metComparator-class}
\alias{.metComparator-class}
\title{\code{.metComparator} class.}
\description{
An S4 class union representing metric condition comparator operators.
}
\keyword{internal}
