% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operand-classes.R
\docType{class}
\name{.metOperand-class}
\alias{.metOperand-class}
\title{\code{.metOperand} class.}
\description{
An S4 class to represent the numerical value of a metric condition expression.
}
\keyword{internal}
