% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Comparator-methods.R
\docType{methods}
\name{not-equals}
\alias{not-equals}
\alias{!=,.var,.operand-method}
\title{Not equals to}
\usage{
\S4method{!=}{.var,.operand}(e1, e2)
}
\arguments{
\item{e1}{Dimension or metric object}

\item{e2}{Operand object}
}
\description{
Expr Do the values on the left and right not match.
}
\seealso{
Other comparator_functions: \code{\link{<,.var,.metOperand-method}},
  \code{\link{<=,.var,.metOperand-method}},
  \code{\link{==,.var,.operand-method}},
  \code{\link{>,.var,.metOperand-method}},
  \code{\link{>=,.var,.metOperand-method}},
  \code{\link{Comparator}}, \code{\link{\%between\%}},
  \code{\link{\%contains\%}}, \code{\link{\%ends_with\%}},
  \code{\link{\%in\%}}, \code{\link{\%matches\%}},
  \code{\link{\%starts_with\%}}
}
\concept{comparator_functions}
