% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Bool.R, R/all-generics.R
\docType{methods}
\name{Not,.operator-method}
\alias{!,.expr-method}
\alias{!,.gaSimpleOrSequence-method}
\alias{!,.operator-method}
\alias{!,orExpr-method}
\alias{Not}
\alias{Not,.expr-method}
\alias{Not,.gaSimpleOrSequence-method}
\alias{Not,.operator-method}
\alias{Not,orExpr-method}
\title{Not.}
\usage{
\S4method{Not}{.operator}(object)

\S4method{!}{.operator}(x)

\S4method{Not}{.expr}(object)

\S4method{!}{.expr}(x)

\S4method{Not}{orExpr}(object)

\S4method{!}{orExpr}(x)

\S4method{Not}{.gaSimpleOrSequence}(object)

\S4method{!}{.gaSimpleOrSequence}(x)

Not(object)
}
\arguments{
\item{object}{An object to get the logical inverse of.}

\item{x}{the object to return the logical inverse of.}
}
\description{
Invert an expression, i.e. NOT.
}
\section{Methods (by class)}{
\itemize{
\item \code{.operator}: 

\item \code{.operator}: 

\item \code{.expr}: 

\item \code{.expr}: 

\item \code{orExpr}: 

\item \code{orExpr}: 

\item \code{.gaSimpleOrSequence}: 

\item \code{.gaSimpleOrSequence}: 
}}

