% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/all-generics.R
\name{Operator}
\alias{Operator}
\alias{Operator<-}
\title{Operator}
\usage{
Operator(object)

Operator(object) <- value
}
\description{
Get or create an operator used in an expression.

Set the operator used in an expression.
}

