% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Segment.R, R/all-generics.R
\docType{methods}
\name{PerHit,gaMetExpr-method}
\alias{PerHit}
\alias{PerHit,gaMetExpr-method}
\title{PerHit}
\usage{
\S4method{PerHit}{gaMetExpr}(object, ...)

PerHit(object, ...)
}
\arguments{
\item{object}{a gaMetExpr object to coerce to user-level.}

\item{...}{Other conditions to be included in the segment definition.}
}
\description{
Set the scope of a gaMetExpr object to hit-level.
}
\section{Methods (by class)}{
\itemize{
\item \code{gaMetExpr}: 
}}

