% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Segment-generics.R, R/Segment-methods.R
\docType{methods}
\name{PerProduct}
\alias{PerProduct}
\alias{PerProduct,gaMetExpr-method}
\alias{PerProduct,formula-method}
\title{PerProduct.}
\usage{
PerProduct(object)

\S4method{PerProduct}{gaMetExpr}(object)

\S4method{PerProduct}{formula}(object)
}
\arguments{
\item{object}{a gaMetExpr object to coerce to hit-level}
}
\value{
a gaMetExpr object.
}
\description{
Set the scope of a gaMetExpr object to product-level.
}
\section{Methods (by class)}{
\itemize{
\item \code{gaMetExpr}: Set the scope of the supplied metric condition to
product-level.

\item \code{formula}: Set the scope of the supplied non-standard-evaluation
metric condition to product-level.
}}

