% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/all-generics.R
\name{GetGaData}
\alias{GetGaData}
\title{GetGaData.}
\usage{
GetGaData(query, creds = NULL, ...)
}
\arguments{
\item{query}{The query execute and returned the processed response for.}

\item{creds}{The OAuth2.0 credentials to use for the request.}

\item{...}{Other arguments to pass on to lower-level API functions.}
}
\description{
Fetch the data for the Google Analytics API query.
}

