% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparator-classes.R
\docType{class}
\name{.rtComparator-class}
\alias{.rtComparator-class}
\title{\code{.rtComparator} class.}
\description{
An S4 class union representing Real-Time condition comparator operators.
}
\keyword{internal}
