% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/comparator-classes.R
\docType{class}
\name{rtDimComparator-class}
\alias{rtDimComparator-class}
\title{`rtDimComparator` class.}
\description{
An S4 class to represent Real-Time dimension condition comparator operators.
}
\keyword{internal}

