% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/var-list-classes.R
\docType{class}
\name{rtDimensions-class}
\alias{rtDimensions-class}
\title{`rtDimensions` class.}
\description{
An S4 class to represent a list of Real-Time reporting dimensions.
}
\keyword{internal}

