% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/all-generics.R
\name{RtDimensions}
\alias{RtDimensions}
\alias{RtDimensions<-}
\title{RtDimensions}
\usage{
RtDimensions(.Object, ...)

RtDimensions(.Object) <- value
}
\description{
Get the dimensions of the object.

Set the dimensions for the object.
}

