% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/expression-classes.R
\docType{class}
\name{rtDimExpr-class}
\alias{rtDimExpr-class}
\title{`rtDimExpr` class.}
\description{
An S4 class to represent Real-Time dimension condition expression.
}
\keyword{internal}

