% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/operand-classes.R
\docType{class}
\name{rtDimOperand-class}
\alias{rtDimOperand-class}
\title{`rtDimOperand` class.}
\description{
An S4 class to represent a Real-Time dimension condition operand.
}
\keyword{internal}

