% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/var-classes.R
\docType{class}
\name{rtDimVar-class}
\alias{rtDimVar-class}
\title{`rtDimVar` class.}
\description{
An S4 class to represent a Real-Time reporting dimension.
}
\details{
Set to "rt:minutesAgo" by default.
}
\keyword{internal}

