% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/comparator-classes.R
\docType{class}
\name{rtMetComparator-class}
\alias{rtMetComparator-class}
\title{`rtMetComparator` class.}
\description{
An S4 class representing Real-Time metric condition comparator operators.
}
\keyword{internal}

