% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/operand-classes.R
\docType{class}
\name{rtMetOperand-class}
\alias{rtMetOperand-class}
\title{`rtMetOperand` class.}
\description{
An S4 class to represent a Real-Time metric condition operand.
}
\keyword{internal}

