% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/all-generics.R
\name{RtMetrics}
\alias{RtMetrics}
\alias{RtMetrics<-}
\title{RtMetrics}
\usage{
RtMetrics(.Object, ...)

RtMetrics(.Object) <- value
}
\description{
Get the metrics of the object.

Set the metrics of the object.
}

