% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/all-generics.R
\name{RtOperand}
\alias{RtOperand}
\alias{RtOperand<-}
\title{RtOperand}
\usage{
RtOperand(.Object)

RtOperand(.Object) <- value
}
\description{
Get the operand of an expression.

Set the operand of an expression.
}

