% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Query.R
\name{RtQuery}
\alias{RtQuery}
\title{RtQuery.}
\usage{
RtQuery(view = NA, creds = .creds, metrics = "rt:pageviews",
  dimensions = "rt:minutesAgo", sortBy = NULL, filters = NULL,
  maxResults = kGaMaxResults)
}
\arguments{
\item{view}{view id to use}

\item{creds}{authentication credentials object created using GoogleApiCreds()}

\item{metrics}{character vector of metrics}

\item{dimensions}{character vector of dimensions}

\item{sortBy}{a sort by object}

\item{filters}{a filters object}

\item{maxResults}{the maximum number of results to return,
 up to 1,000,000}
}
\description{
Create a Real-Time reporting API query object
}

