% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/all-generics.R
\name{RtVar}
\alias{RtVar}
\alias{RtVar<-}
\title{RtVar}
\usage{
RtVar(.Object)

RtVar(.Object) <- value
}
\description{
Gets or Creates an object from the superclass .rtVar

Sets the value of an object or sets its slot belonging to the superclass .rtVar
}

