% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Expr-generics.R, R/Expr-methods.R,
%   R/Segment-methods.R
\docType{methods}
\name{ScopeLevel}
\alias{ScopeLevel}
\alias{ScopeLevel<-}
\alias{ScopeLevel,gaSegMetExpr-method}
\alias{ScopeLevel<-,gaMetExpr,character-method}
\alias{ScopeLevel,gaSegmentFilterList-method}
\alias{ScopeLevel<-,gaSegmentFilterList,character-method}
\title{ScopeLevel.}
\usage{
ScopeLevel(object, value)

ScopeLevel(object) <- value

\S4method{ScopeLevel}{gaSegMetExpr}(object)

\S4method{ScopeLevel}{gaMetExpr,character}(object) <- value

\S4method{ScopeLevel}{gaSegmentFilterList}(object)

\S4method{ScopeLevel}{gaSegmentFilterList,character}(object) <- value
}
\arguments{
\item{object}{a gaSegmentFilterList or a metric expression.}

\item{value}{Optional new scope level to return an updated copy of the object
with the new scope applied.}

\item{object}{a gaSegmentFilterList or a metric expression.}

\item{value}{The scope level to apply. For gaSegmentFilterLists this can be
either 'users' or 'sessions'. For metric expressions use either 'perUser',
'perSession', 'perHit' or 'perProduct'.}
}
\value{
the scope level as a character string, a gaSegmentFilterList or gaMetExpr.
}
\description{
Get the scope level of a gaSegmentFilterList or gaMetExpr.

Set the scope level of a gaDynSegment or a gaMetExpr
}
\section{Methods (by class)}{
\itemize{
\item \code{gaSegMetExpr}: Return the scope of the supplied metric used within a
segment definition.

\item \code{object = gaMetExpr,value = character}: Set the scope, as described by a character value, to
be applied to the supplied metric condition for use within a segment
expression.

\item \code{gaSegmentFilterList}: Return the scope of the supplied gaSegmentFilterList.

\item \code{object = gaSegmentFilterList,value = character}: Set the scope level of a gaSegmentFilterList to either
"user" or "session" level.
}}

