% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Segment-generics.R, R/Segment-methods.R
\docType{methods}
\name{SegmentFilters}
\alias{SegmentFilters}
\alias{SegmentFilters,ANY,ANY-method}
\alias{SegmentFilters,gaSegmentFilterList,missing-method}
\title{SegmentFilters.}
\usage{
SegmentFilters(object, ..., scope = "sessions")

\S4method{SegmentFilters}{ANY,ANY}(object, ..., scope = "sessions")

\S4method{SegmentFilters}{gaSegmentFilterList,missing}(object)
}
\arguments{
\item{object}{The first filter to include in the segment definition.}

\item{...}{Additional filters to be included in the segment definition.}

\item{scope}{The scope of the resulting gaSegmentFilterList, either 'user' or
'session' level.}
}
\value{
a gaSegmentFilterList object.
}
\description{
Combine one or more segment condition filters and/or sequence filters into
a gaSegmentFilterList that is scoped to either 'user' or 'session' level.
}
\details{
A segment filter is either sequential or non-sequential conditions. Sequential
and non-sequential conditoins can be combined using this function.
}
\section{Methods (by class)}{
\itemize{
\item \code{object = ANY,scope = ANY}: Define a list of filters from one or more
expressions applied using the specified scope.

\item \code{object = gaSegmentFilterList,scope = missing}: Returns itself.
}}

