% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sequence.R
\name{sequential_segment}
\alias{sequential_segment}
\title{sequential_segment.}
\usage{
sequential_segment(steps)
}
\arguments{
\item{steps}{a list of expressions (of one or more conditions), each
  repreating a step in the sequence.}
}
\description{
Create a sequence using non-standard evaluation.
}
\details{
Steps must be separated by commas (\code{,}). ... denotes zero or more of any
interaction preceeding the step that follows.
}
\examples{
a <- condition(pagePath == "/home")
b <- condition(eventCategory == "Video") &
  condition(eventAction == "Play")
c <- condition(medium == "email")
s <- sequential_segment(list( ..., a, ..., b, c ))
}

