% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Segment-methods.R
\name{setSegmentFilterScopeNegation}
\alias{setSegmentFilterScopeNegation}
\title{setSegmentFilterScopeNegation}
\usage{
setSegmentFilterScopeNegation(object, negation, scope)
}
\arguments{
\item{object}{a \code{.gaSegmentFilter} object, or an object that can be
coerced to a \code{.gaSegmentFilter} object.}

\item{negation}{optional boolean value of \code{TRUE} or \code{FALSE}.}

\item{scope}{optional either \code{"sessions"} or \code{"users"}.}
}
\value{
a \code{.gaSegmentFilter} object.
}
\description{
Given a \code{.gaSegmentFilter} object, or an object that can be coerced to a
\code{.gaSegmentFilter} object, set the \code{negation} and/or \code{scope}
of that \code{.gaSegmentFilter} object as per the non-missing arguments.
}
\keyword{internal}
