% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/all-generics.R
\name{SortBy}
\alias{SortBy}
\alias{SortBy<-}
\title{SortBy}
\usage{
SortBy(.Object, ..., desc = logical(0))

SortBy(.Object) <- value
}
\description{
Get the sortBy order of the query.

Set the order of rows returned by Google Analytics.
}

