% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{split_permissions}
\alias{split_permissions}
\title{split_permissions}
\usage{
split_permissions(permissions)
}
\description{
Take a list of character vector describing the permissions for each user and
transform into a nested list of users and their list of permissions.
}
\keyword{internal}

