% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GaDate.R
\name{SplitDateRange}
\alias{SplitDateRange}
\title{SplitDateRange
Splits a gaDateRange object into N pieces. Useful for splitting a query into
smaller chunks in order to overcome sampling.}
\usage{
SplitDateRange(dateRange, N)
}
\arguments{
\item{dateRange}{the gaDateRange object to be split}

\item{N}{the number of the separate date ranges to be split into; use 0 for
single days.}
}
\description{
SplitDateRange
Splits a gaDateRange object into N pieces. Useful for splitting a query into
smaller chunks in order to overcome sampling.
}

