% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query-classes.R
\docType{class}
\name{.standardQuery-class}
\alias{.standardQuery-class}
\title{\code{.standardQuery} class.}
\description{
An S4 class to represent a standard reporting API query.
}
\section{Slots}{

\describe{
\item{\code{dateRange}}{The date range for the query as a dateRange object}

\item{\code{samplingLevel}}{The sampling level of the query as a character string.}
}}

\keyword{internal}
