% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Date-generics.R, R/Date-methods.R
\name{StartDate}
\alias{StartDate}
\alias{StartDate<-}
\alias{StartDate,character-method}
\alias{StartDate,dateRange-method}
\alias{StartDate,Interval-method}
\alias{StartDate,.standardQuery-method}
\alias{StartDate,gaView-method}
\alias{StartDate<-,dateRange-method}
\alias{StartDate<-,.standardQuery-method}
\title{StartDate}
\usage{
StartDate(object, value)

StartDate(object) <- value

\S4method{StartDate}{character}(object)

\S4method{StartDate}{dateRange}(object)

\S4method{StartDate}{Interval}(object)

\S4method{StartDate}{.standardQuery}(object)

\S4method{StartDate}{gaView}(object)

\S4method{StartDate}{dateRange}(object) <- value

\S4method{StartDate}{.standardQuery}(object) <- value
}
\arguments{
\item{object}{Object to get start date of.}

\item{value}{Value to set start date of object to.}
}
\description{
Get the start date.

Set the start date.
}
\section{Functions}{
\itemize{
\item \code{StartDate,character-method}: Coerce a character vector into a Google Analytics date
object.

\item \code{StartDate,dateRange-method}: Return the start dates of a date range vector.

\item \code{StartDate,Interval-method}: Return the start dates of a date range vector.

\item \code{StartDate,.standardQuery-method}: Return the start dates of a query's date range vector.

\item \code{StartDate,gaView-method}: Get the date when a view first started receiving hits.

\item \code{StartDate<-,dateRange-method}: Set a new start date for a date range.

\item \code{StartDate<-,.standardQuery-method}: Set a new start date for a query.
}}

\seealso{
Other date range functions: 
\code{\link{Cohort}()},
\code{\link{DateRange}()},
\code{\link{EndDate}()},
\code{\link{SplitDateRange}()}
}
\concept{date range functions}
