% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/table-filter-classes.R
\docType{class}
\name{.tableFilter-class}
\alias{.tableFilter-class}
\title{`.tableFilter` class.}
\description{
An S4 class to represent a query table filter epxression.
}
\keyword{internal}

