% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TableFilter-generics.R, R/TableFilter-methods.R
\name{TableFilter}
\alias{TableFilter}
\alias{TableFilter<-}
\alias{TableFilter,ANY,missing-method}
\alias{TableFilter,.query,missing-method}
\alias{TableFilter,.query,ANY-method}
\alias{TableFilter<-,.query-method}
\title{TableFilter}
\usage{
TableFilter(object, value)

TableFilter(object) <- value

\S4method{TableFilter}{ANY,missing}(object)

\S4method{TableFilter}{.query,missing}(object)

\S4method{TableFilter}{.query,ANY}(object, value)

\S4method{TableFilter}{.query}(object) <- value
}
\arguments{
\item{object}{The object to be coerced to a TableFilter or the query object
to apply a table filter to.}

\item{value}{The replacement table filter where \code{object} is a query.}
}
\description{
Get the filter.

Set the filter.
}
\section{Functions}{
\itemize{
\item \code{TableFilter,ANY,missing-method}: Coerce the given object into a table filter.

\item \code{TableFilter,.query,missing-method}: Return the TableFilter that has been applied to the
given query.

\item \code{TableFilter,.query,ANY-method}: Method to replace the table filter of a query

\item \code{TableFilter<-,.query-method}: Method to replace the table filter of a query
}}

\seealso{
Other query object functions: 
\code{\link{Dimensions}()},
\code{\link{MaxResults}()},
\code{\link{Metrics}()},
\code{\link{SamplingLevel}()},
\code{\link{SortBy}()}
}
\concept{query object functions}
