% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sequence-generics.R
\name{Then}
\alias{Then}
\title{Then}
\usage{
Then(object, ...)
}
\arguments{
\item{object}{The expression that should \bold{immediately} precede others in
the sequence.}

\item{...}{Any other expressions that should \bold{immediately} follow the
first one but before any others in the sequence.}
}
\value{
A \code{gaSegmentSequenceStep} object, with the immediate flag set.
}
\description{
Treat a step within a sequence as happening immediately after any preceding
steps in the sequence, i.e. 'immediately following'.
}
\examples{
purchased_immediately_after <- Then(Expr(~transactions > 0))

}
\seealso{
\code{\link{Sequence}}

Other sequence segment functions: \code{\link{First}},
  \code{\link{Later}}, \code{\link{Sequence}}
}
\concept{sequence segment functions}
