% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{unnest_objects}
\alias{unnest_objects}
\title{unnest_objects}
\usage{
unnest_objects(..., class)
}
\arguments{
\item{...}{objects to be concatenated, including lists of objects.}

\item{class}{the class of the objects considered as nested.}
}
\description{
Concatenate objects of class \code{class} with lists of objects into a flattened list.
}
\keyword{internal}
