% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{unsplit_permissions}
\alias{unsplit_permissions}
\title{unsplit_permissions}
\usage{
unsplit_permissions(permissions)
}
\description{
Take a list of Google Analytics user permissions for a list of one or more users and
transform the into a list of character vector.
}
\keyword{internal}

