% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/var-classes.R
\docType{class}
\name{.var-class}
\alias{.var-class}
\title{`.var` class.}
\description{
An S4 class to represent the name of a single Google Analytics dimension or
metric variable.
}
\keyword{internal}

