% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Bool.R, R/all-generics.R
\docType{methods}
\name{xor,.compoundExpr,.compoundExpr-method}
\alias{xor}
\alias{xor,.compoundExpr,.compoundExpr-method}
\title{Generate an expression that gives the exclusive or of two expressions.}
\usage{
\S4method{xor}{.compoundExpr,.compoundExpr}(x, y)

xor(x, y)
}
\arguments{
\item{x}{The first expression}

\item{y}{The second expression.}
}
\description{
Generate an expression that gives the exclusive or of two expressions.
}
\section{Methods (by class)}{
\itemize{
\item \code{x = .compoundExpr,y = .compoundExpr}: 
}}

