% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/all-generics.R
\name{And}
\alias{And}
\title{And}
\usage{
And(object, ...)
}
\description{
AND two or more ganalytics expressions together.
}
\details{
Create a new AND expression from one or more arguments
Valid types are either AND, OR, or single expressions.
A single list of objects is also accepted.
}

