% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Bool-generics.R, R/Bool.R
\docType{methods}
\name{And}
\alias{&,.compoundExpr,.compoundExpr-method}
\alias{And}
\alias{And,.compoundExpr-method}
\title{And.}
\usage{
And(object, ...)

\S4method{And}{.compoundExpr}(object, ...)

\S4method{&}{.compoundExpr,.compoundExpr}(e1, e2)
}
\arguments{
\item{object}{The first object within the AND expression}

\item{...}{Additional objects to include within the AND expression.}

\item{e1}{first expression}

\item{e2}{second expression}
}
\value{
an object of class \code{andExpr}
}
\description{
AND two or more ganalytics expressions together.
}
\details{
Create a new AND expression from one or more arguments
Valid types are either AND, OR, or single expressions.
A single list of objects is also accepted.
}
\section{Methods (by class)}{
\itemize{
\item \code{.compoundExpr}: 

\item \code{e1 = .compoundExpr,e2 = .compoundExpr}: 
}}
\seealso{
\code{link{Or}} \code{link{xor}} \code{link{Not}}
}

