% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/helper-functions.R
\name{ArgList}
\alias{ArgList}
\title{ArgList
If the only argument passed was already a list, then extract that list.}
\usage{
ArgList(...)
}
\arguments{
\item{...}{arguments or list of arguments}
}
\description{
ArgList
If the only argument passed was already a list, then extract that list.
}

