% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{ArgList}
\alias{ArgList}
\title{ArgList.}
\usage{
ArgList(...)
}
\arguments{
\item{...}{arguments or list of arguments}
}
\description{
If the only argument passed was already a list, then extract that list.
}
\keyword{internal}

