% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{checkDataFrameClasses}
\alias{checkDataFrameClasses}
\title{checkDataFrameClasses}
\usage{
checkDataFrameClasses(object, matchClasses)
}
\description{
Test whether the class for each column of a data.frame match a list of
expected classes.
}
\keyword{internal}

