% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/helper-functions.R
\name{CheckVectorBounds}
\alias{CheckVectorBounds}
\title{CheckVectorBounds
Check the length of each named slot within object is within the lower and
upper bounds specified.}
\usage{
CheckVectorBounds(object, slot_vector_bound_list)
}
\arguments{
\item{object}{an object with slots that match the names of slot_vector_bound_list}

\item{slot_vector_bound_list}{a named list of vectors specifying the upper
and lower bounds for the length of each slot of object.}
}
\description{
CheckVectorBounds
Check the length of each named slot within object is within the lower and
upper bounds specified.
}

