% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Date-generics.R
\name{Cohort}
\alias{Cohort}
\alias{Cohort<-}
\title{Cohort.}
\usage{
Cohort(object, value, type)

Cohort(object) <- value
}
\arguments{
\item{object}{to get or set the cohorts of.}

\item{value}{to set the object's cohorts to.}

\item{type}{of cohort. Only the default of "FIRST_VISIT_DATE" is currently valid.}
}
\description{
Get or define a cohort.

Set a cohort of a query.
}
\seealso{
Other date range functions: \code{\link{DateRange}},
  \code{\link{EndDate}}, \code{\link{StartDate}}

Other date range functions: \code{\link{DateRange}},
  \code{\link{EndDate}}, \code{\link{StartDate}}
}
\concept{date range functions}
