% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/comparator-classes.R
\docType{class}
\name{.comparator-class}
\alias{.comparator-class}
\title{`.comparator` class.}
\description{
An S4 class to represent a condition's comparator operator.
}
\details{
Set to "==" (exact match) by default.
}
\keyword{internal}

