% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Comparator-generics.R, R/Comparator-methods.R
\docType{methods}
\name{Comparator}
\alias{Comparator}
\alias{Comparator<-}
\alias{Comparator,.expr-method}
\alias{Comparator<-,.expr-method}
\title{Comparator}
\usage{
Comparator(object, ...)

Comparator(object) <- value

\S4method{Comparator}{.expr}(object)

\S4method{Comparator}{.expr}(object) <- value
}
\arguments{
\item{object}{The object to be coerced to a '.comparator' subclass or the
expression object of which to obtain its comparator.}

\item{...}{Used by certain methods.}

\item{value}{The value to set the comparator to.}
}
\value{
An object belonging to the superclass \code{.comparator}
}
\description{
Get the comparator used in an expression or create a comparator object.

Set the comparator of an expression.
}
\section{Methods (by class)}{
\itemize{
\item \code{.expr}: Return the comparator used within the supplied
conditional expression.

\item \code{.expr}: Replace the comparator of the supplied
conditional expression.
}}

\seealso{
Other comparator functions: \code{\link{comparators}}

Other comparator functions: \code{\link{comparators}}
}
\concept{comparator functions}
