% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/expression-classes.R
\docType{class}
\name{.compoundExpr-class}
\alias{.compoundExpr-class}
\title{`.compoundExpr` class.}
\description{
An S4 class representing an expression containing one or more conditions.
}
\keyword{internal}

