% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AutoExpr.R
\name{condition}
\alias{condition}
\title{condition}
\usage{
condition(cond)
}
\description{
Define a Google Analytics condition using a domain-specific language familiar to R users and using non-standard evaluation.
}

