% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Date.R, R/all-generics.R
\docType{methods}
\name{StartDate,character-method}
\alias{DateRange}
\alias{DateRange,.standardQuery,ANY-method}
\alias{DateRange,Date,Date-method}
\alias{DateRange,character,character-method}
\alias{DateRange,dateRange,ANY-method}
\alias{DateRange<-}
\alias{DateRange<-,.standardQuery,.standardQuery-method}
\alias{DateRange<-,.standardQuery,Date-method}
\alias{DateRange<-,.standardQuery,character-method}
\alias{DateRange<-,.standardQuery,dateRange-method}
\alias{DateRange<-,dateRange,.standardQuery-method}
\alias{DateRange<-,dateRange,character-method}
\alias{DateRange<-,dateRange,dateRange-method}
\alias{EndDate}
\alias{EndDate,.standardQuery-method}
\alias{EndDate,character-method}
\alias{EndDate,dateRange-method}
\alias{EndDate<-}
\alias{EndDate<-,.standardQuery,Date-method}
\alias{EndDate<-,.standardQuery,character-method}
\alias{EndDate<-,dateRange,Date-method}
\alias{EndDate<-,dateRange,character-method}
\alias{StartDate}
\alias{StartDate,.standardQuery-method}
\alias{StartDate,character-method}
\alias{StartDate,dateRange-method}
\alias{StartDate<-}
\alias{StartDate<-,.standardQuery,Date-method}
\alias{StartDate<-,.standardQuery,character-method}
\alias{StartDate<-,dateRange,Date-method}
\alias{StartDate<-,dateRange,character-method}
\title{DateRange.}
\usage{
\S4method{StartDate}{character}(object)

\S4method{EndDate}{character}(object)

\S4method{StartDate}{dateRange}(object)

\S4method{EndDate}{dateRange}(object)

\S4method{StartDate}{.standardQuery}(object)

\S4method{EndDate}{.standardQuery}(object)

\S4method{StartDate}{dateRange,character}(object) <- value

\S4method{EndDate}{dateRange,character}(object) <- value

\S4method{StartDate}{dateRange,Date}(object) <- value

\S4method{EndDate}{dateRange,Date}(object) <- value

\S4method{StartDate}{.standardQuery,character}(object) <- value

\S4method{EndDate}{.standardQuery,character}(object) <- value

\S4method{StartDate}{.standardQuery,Date}(object) <- value

\S4method{EndDate}{.standardQuery,Date}(object) <- value

\S4method{DateRange}{dateRange,ANY}(object)

\S4method{DateRange}{character,character}(object, endDate)

\S4method{DateRange}{Date,Date}(object, endDate)

\S4method{DateRange}{.standardQuery,ANY}(object)

\S4method{DateRange}{dateRange,character}(object) <- value

\S4method{DateRange}{.standardQuery,dateRange}(object) <- value

\S4method{DateRange}{.standardQuery,.standardQuery}(object) <- value

\S4method{DateRange}{dateRange,dateRange}(object) <- value

\S4method{DateRange}{dateRange,.standardQuery}(object) <- value

\S4method{DateRange}{.standardQuery,character}(object) <- value

\S4method{DateRange}{.standardQuery,Date}(object) <- value

DateRange(object, endDate)

DateRange(object) <- value

StartDate(object, ...)

StartDate(object) <- value

EndDate(object, ...)

EndDate(object) <- value
}
\description{
Get the date range.

Set the date range.

Get the start date.

Set the start date.

Get the end date of the date range.

Set the endDate of the date range.
}
\section{Methods (by class)}{
\itemize{
\item \code{character}: 

\item \code{character}: 

\item \code{dateRange}: 

\item \code{dateRange}: 

\item \code{.standardQuery}: 

\item \code{.standardQuery}: 

\item \code{object = dateRange,value = character}: 

\item \code{object = dateRange,value = character}: 

\item \code{object = dateRange,value = Date}: 

\item \code{object = dateRange,value = Date}: 

\item \code{object = .standardQuery,value = character}: 

\item \code{object = .standardQuery,value = character}: 

\item \code{object = .standardQuery,value = Date}: 

\item \code{object = .standardQuery,value = Date}: 

\item \code{object = dateRange,endDate = ANY}: 

\item \code{object = character,endDate = character}: 

\item \code{object = Date,endDate = Date}: 

\item \code{object = .standardQuery,endDate = ANY}: 

\item \code{object = dateRange,value = character}: 

\item \code{object = .standardQuery,value = dateRange}: 

\item \code{object = .standardQuery,value = .standardQuery}: 

\item \code{object = dateRange,value = dateRange}: 

\item \code{object = dateRange,value = .standardQuery}: 

\item \code{object = .standardQuery,value = character}: 

\item \code{object = .standardQuery,value = Date}: 
}}

