% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/comparator-classes.R
\docType{class}
\name{.dimComparator-class}
\alias{.dimComparator-class}
\title{`.dimComparator` class.}
\description{
An S4 class union representing dimension condition comparator operators.
}
\keyword{internal}

