% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Var-list-generics.R, R/Var-methods.R
\docType{methods}
\name{Dimensions}
\alias{Dimensions}
\alias{Dimensions,.query-method}
\alias{Dimensions,ANY-method}
\alias{Dimensions<-}
\alias{Dimensions<-,.query-method}
\title{Dimensions.}
\usage{
Dimensions(object, ...)

Dimensions(object) <- value

\S4method{Dimensions}{ANY}(object, ...)

\S4method{Dimensions}{.query}(object)

\S4method{Dimensions}{.query}(object) <- value
}
\arguments{
\item{object}{An object to be coerced to a list of dimensions.}

\item{...}{Other dimensions to add to the returned list, or if \code{object}
is a query object, the replacement dimensions.}

\item{value}{The replacement dimensions for the supplied object.}
}
\description{
Get the dimensions of the object.

Set the dimensions for the object.
}
\section{Methods (by class)}{
\itemize{
\item \code{ANY}: Coerces the supplied character vector or list into a vector of
Google Analytics dimensions.

\item \code{.query}: Returns the dimensions used within the supplied query.

\item \code{.query}: Replace the dimensions of the query.
}}

