% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Var-list-generics.R, R/Var-list-methods.R
\name{Dimensions}
\alias{Dimensions}
\alias{Dimensions<-}
\alias{Dimensions,ANY-method}
\alias{Dimensions,.query-method}
\alias{Dimensions<-,.query-method}
\title{Dimensions}
\usage{
Dimensions(object, ...)

Dimensions(object) <- value

\S4method{Dimensions}{ANY}(object, ...)

\S4method{Dimensions}{.query}(object)

\S4method{Dimensions}{.query}(object) <- value
}
\arguments{
\item{object}{An object to be coerced to a list of dimensions.}

\item{...}{Other dimensions to add to the returned list, or if \code{object}
is a query object, the replacement dimensions.}

\item{value}{The replacement dimensions for the supplied object.}
}
\description{
Get the dimensions of the object.

Set the dimensions for the object.
}
\section{Functions}{
\itemize{
\item \code{Dimensions,ANY-method}: Coerces the supplied character vector or list into a vector of
Google Analytics dimensions.

\item \code{Dimensions,.query-method}: Returns the dimensions used within the supplied query.

\item \code{Dimensions<-,.query-method}: Replace the dimensions of the query.
}}

\seealso{
Other query object functions: 
\code{\link{MaxResults}()},
\code{\link{Metrics}()},
\code{\link{SamplingLevel}()},
\code{\link{SortBy}()},
\code{\link{TableFilter}()}

Other query object functions: 
\code{\link{MaxResults}()},
\code{\link{Metrics}()},
\code{\link{SamplingLevel}()},
\code{\link{SortBy}()},
\code{\link{TableFilter}()}
}
\concept{query object functions}
