% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/all-generics.R
\name{Dimensions}
\alias{Dimensions}
\alias{Dimensions<-}
\title{Dimensions}
\usage{
Dimensions(object, ...)

Dimensions(object) <- value
}
\description{
Get the dimensions of the object.

Set the dimensions for the object.
}

