% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Var-list-generics.R, R/Var-methods.R
\docType{methods}
\name{Dimensions}
\alias{Dimensions}
\alias{Dimensions,.dimensions-method}
\alias{Dimensions,.query-method}
\alias{Dimensions,NULL-method}
\alias{Dimensions,character-method}
\alias{Dimensions,list-method}
\alias{Dimensions<-}
\alias{Dimensions<-,.query-method}
\title{Dimensions.}
\usage{
Dimensions(object, ...)

Dimensions(object) <- value

\S4method{Dimensions}{`NULL`}(object, ...)

\S4method{Dimensions}{.dimensions}(object, ...)

\S4method{Dimensions}{character}(object, ...)

\S4method{Dimensions}{list}(object, ...)

\S4method{Dimensions}{.query}(object)

\S4method{Dimensions}{.query}(object) <- value
}
\arguments{
\item{object}{An object to be coerced to a list of dimensions.}

\item{...}{Other dimensions to add to the returned list, or if \code{object}
  is a query object, the replacement dimensions.}

\item{value}{The replacement dimensions for the supplied object.}
}
\description{
Get the dimensions of the object.

Set the dimensions for the object.
}
\section{Methods (by class)}{
\itemize{
\item \code{NULL}: 

\item \code{.dimensions}: 

\item \code{character}: 

\item \code{list}: 

\item \code{.query}: 

\item \code{.query}: 
}}

