% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Date-generics.R
\name{EndDate}
\alias{EndDate}
\alias{EndDate<-}
\title{EndDate}
\usage{
EndDate(object, value)

EndDate(object) <- value
}
\arguments{
\item{object}{Object to get end date of.}

\item{value}{Value to set end date of object to.}
}
\description{
Get the end date of the date range.

Set the endDate of the date range.
}
\seealso{
Other date range functions: \code{\link{Cohort}},
  \code{\link{DateRange}}, \code{\link{SplitDateRange}},
  \code{\link{StartDate}}

Other date range functions: \code{\link{Cohort}},
  \code{\link{DateRange}}, \code{\link{SplitDateRange}},
  \code{\link{StartDate}}
}
\concept{date range functions}
