% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/all-generics.R
\name{EndDate}
\alias{EndDate}
\alias{EndDate<-}
\title{EndDate}
\usage{
EndDate(object)

EndDate(object) <- value
}
\description{
Get the end date of the date range.

Set the endDate of the date range.
}

