% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Segment-generics.R, R/Segment-methods.R
\docType{methods}
\name{Exclude}
\alias{Exclude}
\alias{Exclude,ANY-method}
\title{Exclude.}
\usage{
Exclude(object, ..., scope)

\S4method{Exclude}{ANY}(object, ..., scope)
}
\arguments{
\item{object}{a segment condition or sequence filter to exclude.}

\item{...}{additional segment conditions to add to the exclude filter.}

\item{scope}{optional scope, "users" or "sessions"}
}
\value{
a .gaSegmentFilter object with its negate slot set to TRUE.
}
\description{
Set the negation flag of a segment filter to TRUE.
}
\section{Methods (by class)}{
\itemize{
\item \code{ANY}: Define an exclude segment filter using the supplied
expressions.
}}

