% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Segment-generics.R, R/Segment-methods.R
\docType{methods}
\name{Exclude}
\alias{Exclude}
\alias{Exclude,ANY-method}
\title{Exclude.}
\usage{
Exclude(object)

\S4method{Exclude}{ANY}(object)
}
\arguments{
\item{object}{a segment filter condition or sequence to exclude.}
}
\value{
a .gaSegmentFilter object with its negate slot set to TRUE.
}
\description{
Define a segment filter with its negation flag is set to TRUE.
}
\section{Methods (by class)}{
\itemize{
\item \code{ANY}: Define an exclude segment filter using the supplied
expressions.
}}

