% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Segment-generics.R, R/Segment-methods.R
\docType{methods}
\name{Exclude}
\alias{Exclude}
\alias{Exclude,.compoundExpr-method}
\alias{Exclude,gaSegmentSequenceFilter-method}
\title{Exclude.}
\usage{
Exclude(object, ..., scope = "sessions")

\S4method{Exclude}{.compoundExpr}(object, ..., scope = "sessions")

\S4method{Exclude}{gaSegmentSequenceFilter}(object, ..., scope = "sessions")
}
\arguments{
\item{object}{a condition or sequence to exclude}

\item{...}{further conditions or sequences to exclude.}

\item{scope}{the scope of the returned gaSegmentList, either "users" OR
"sessions".}
}
\value{
a .gaSegmentFilter object with its negate slot set to TRUE.
}
\description{
One or more segment conditions or sequences to exclude from the defined user
or session segment.
}
\section{Methods (by class)}{
\itemize{
\item \code{.compoundExpr}: Define an exclude segment filter using the supplied
expressions.

\item \code{gaSegmentSequenceFilter}: Set the negatation field of the supplied sequence filter
so as to exclude matching sequences.
}}

