% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Expr.R, R/all-generics.R
\docType{methods}
\name{Expr,.expr,ANY-method}
\alias{Expr}
\alias{Expr,.expr,ANY-method}
\alias{Expr,character,character-method}
\alias{Expr,formula,ANY-method}
\title{Expr.}
\usage{
\S4method{Expr}{.expr,ANY}(object)

\S4method{Expr}{formula,ANY}(object)

\S4method{Expr}{character,character}(object, comparator, operand,
  metricScope = "")

Expr(object, comparator, operand, metricScope = "")
}
\arguments{
\item{object}{A dimension or metric variable, or another object to be coerced
to an .expr object.}

\item{comparator}{The comparator to use for the expression.}

\item{operand}{The operand to use for the expression.}

\item{metricScope}{The scope to use for segmentation if using a metric.
  Possible values include "perUser" or "perSession".}
}
\description{
Define a Google Analytics expression.
}
\section{Methods (by class)}{
\itemize{
\item \code{object = .expr,comparator = ANY}: Accepts a formula in the form of: \code{~ <variable> <comparator>
<operand>} where only the \code{<operand>} is evaluated.

\item \code{object = formula,comparator = ANY}: 

\item \code{object = character,comparator = character}: 
}}
\seealso{
Other expression generators: \code{\link{GaExpr}},
  \code{\link{GaExpr,character,character-method}};
  \code{\link{McfExpr}},
  \code{\link{McfExpr,character,character-method}};
  \code{\link{RtExpr}},
  \code{\link{RtExpr,character,character-method}}
}

