% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{flatten}
\alias{flatten}
\title{flatten.}
\usage{
flatten(x)
}
\arguments{
\item{x}{a list type object to flatten.}
}
\value{
a list
}
\description{
Flatten a nested list while preserving the class of each element
Convert a list type object into a non-nested list, preserving
the original object classes.
}
\references{
\url{http://stackoverflow.com/a/8139959/1007029}
}
\keyword{internal}

