% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/management-api-classes.R
\name{GaAccountSummary}
\alias{GaAccountSummary}
\title{GA Account Summary}
\usage{
GaAccountSummary(id = NULL, creds = .creds)
}
\arguments{
\item{id}{ID of the GA account to get a summary of.}

\item{creds}{The Google APIs credentials to use.}
}
\description{
Get a Google Analytics account summary resource.
}

