% Generated by roxygen2 (4.0.2): do not edit by hand
\name{GaAnd}
\alias{GaAnd}
\title{GaAnd}
\usage{
GaAnd(.Object, ...)
}
\description{
AND two or more ganalytics expressions together.
}
\details{
Create a new AND expression from one or more arguments
Valid types are either AND, OR, or single expressions.
A single list of objects is also accepted.
}

