% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/deprecated.R
\name{GaAnd}
\alias{GaAnd}
\title{GaAnd (Deprecated, use And or & instead).}
\usage{
GaAnd(...)
}
\arguments{
\item{...}{arguments passed to \code{And}}
}
\description{
GaAnd (Deprecated, use And or & instead).
}

