% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/deprecated.R
\name{GaCondition}
\alias{GaCondition}
\title{GaCondition}
\usage{
GaCondition(..., negation = FALSE)
}
\arguments{
\item{...}{passed onto \code{Include} or \code{Exclude}}

\item{negation}{used to select either Include or Exclude (if negation is
  FALSE).}
}
\description{
Deprecated, use \code{Include} or \code{Exclude} instead.
}
\keyword{internal}

