% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Creds.R, R/all-generics.R
\docType{methods}
\name{GaCreds,.query,list-method}
\alias{GaCreds}
\alias{GaCreds,.query,ANY-method}
\alias{GaCreds,.query,list-method}
\alias{GaCreds,character,ANY-method}
\alias{GaCreds,missing,ANY-method}
\alias{GaCreds<-}
\alias{GaCreds<-,.query,list-method}
\title{Authentication credentials for Google Analytics API queries.}
\usage{
\S4method{GaCreds}{.query,list}(object, value)

\S4method{GaCreds}{.query,ANY}(object)

\S4method{GaCreds}{character,ANY}(object = "GANALYTICS", value = NULL, ...)

\S4method{GaCreds}{missing,ANY}(object = "GANALYTICS", value = NULL, ...)

\S4method{GaCreds}{.query,list}(object) <- value

GaCreds(object = "GANALYTICS", value = NULL, ...)

GaCreds(object) <- value
}
\arguments{
\item{object}{The object to get the credentials from.}

\item{value}{The replacement credentials for the supplied query object.}

\item{...}{other arguments pass to \code{GoogleApiCreds}.}
}
\description{
Get or set the authentication credentials for a Google Analytics query object.

Set the authentication credentials for a Google Analytics query object.
}
\section{Methods (by class)}{
\itemize{
\item \code{object = .query,value = list}: 

\item \code{object = .query,value = ANY}: 

\item \code{object = character,value = ANY}: 

\item \code{object = missing,value = ANY}: 

\item \code{object = .query,value = list}: 
}}

