% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/deprecated.R
\name{GaDateRange}
\alias{GaDateRange}
\alias{GaDateRange<-}
\title{GaDateRange}
\usage{
GaDateRange(...)

GaDateRange(...) <- value
}
\arguments{
\item{...}{arguments passed onto \code{DateRange}}

\item{value}{passed onto \code{DateRange}}
}
\description{
Deprecated, use DateRange instead.

Deprecated, use DateRange instead.
}
\keyword{internal}

