% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/all-generics.R
\name{GaDimensions}
\alias{GaDimensions}
\alias{GaDimensions<-}
\title{GaDimensions}
\usage{
GaDimensions(.Object, ...)

GaDimensions(.Object) <- value
}
\description{
Get the dimensions of the object.

Set the dimensions for the object.
}

