% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/segment-classes.R
\docType{class}
\name{gaDynSegment-class}
\alias{gaDynSegment-class}
\title{`gaDynSegment` class.}
\description{
An S4 class to represent a list of scoped segment filter lists.
}
\keyword{internal}

