% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/backwards-compatibility.R
\name{GaEndDate}
\alias{GaEndDate}
\alias{GaEndDate<-}
\title{GaEndDate (Deprecated)}
\usage{
GaEndDate(...)

GaEndDate(...) <- value
}
\arguments{
\item{...}{arguments passed onto \code{EndDate}}

\item{value}{passed onto \code{EndDate}}
}
\description{
GaEndDate (Deprecated)

GaEndDate<- (Deprecated)
}

