% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/deprecated.R
\name{GaEndDate}
\alias{GaEndDate}
\alias{GaEndDate<-}
\title{GaEndDate}
\usage{
GaEndDate(...)

GaEndDate(...) <- value
}
\arguments{
\item{...}{arguments passed onto \code{EndDate}}

\item{value}{passed onto \code{EndDate}}
}
\description{
Deprecated, use EndDate instead.

Deprecated, use EndDate instead.
}
\keyword{internal}

