% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/all-generics.R
\name{GaExpr}
\alias{GaExpr}
\title{GaExpr}
\usage{
GaExpr(.Object, operator, operand, metricScope = "")
}
\description{
Create an expression.
}
\examples{
\donttest{
  myQuery <- GaQuery(profileId = 123456789)
  source_matches_google <- GaExpr("source", "~", "google")
  GaFilter(myQuery) <- source_matches_google
  GetGaData(myQuery)
}
}

