% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/all-generics.R
\name{Expr}
\alias{Expr}
\alias{GaExpr}
\title{Expr}
\usage{
Expr(.Object, operator, operand, metricScope = "")

GaExpr(.Object, operator, operand, metricScope = "")
}
\description{
Expr

Create an expression.
}
\examples{
\donttest{
  myQuery <- GaQuery(view = 123456789)
  source_matches_google <- GaExpr("source", "~", "google")
  GaFilter(myQuery) <- source_matches_google
  GetGaData(myQuery)
}
}

