% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Expr.R, R/all-generics.R
\docType{methods}
\name{GaExpr,character,character-method}
\alias{GaExpr}
\alias{GaExpr,character,character-method}
\title{GaExpr.}
\usage{
\S4method{GaExpr}{character,character}(object, operator, operand,
  metricScope = "")

GaExpr(object, operator, operand, metricScope = "")
}
\arguments{
\item{object}{A dimension or metric variable, or another object to be coerced
to an .expr object.}

\item{operator}{The operator to use for the expression.}

\item{operand}{The operand to use for the expression.}

\item{metricScope}{The scope to use for segmentation if using a metric.
  Possible values include "perUser" or "perSession".}
}
\description{
Create a Core Reporting API expression.
}
\section{Methods (by class)}{
\itemize{
\item \code{object = character,operator = character}: 
}}
\examples{
\donttest{
  myQuery <- GaQuery(view = 123456789)
  source_matches_google <- GaExpr("source", "~", "google")
  GaFilter(myQuery) <- source_matches_google
  GetGaData(myQuery)
}
}

