% Generated by roxygen2 (4.0.1): do not edit by hand
\name{GaExpr}
\alias{GaExpr}
\title{GaExpr}
\usage{
GaExpr(.Object, gaOperator, gaOperand)
}
\description{
Create an expression.
}
\examples{
\donttest{
  myQuery <- GaQuery(profileId = 123456789)
  source_matches_google <- GaExpr("source", "~", "google")
  GaFilter(myQuery) <- source_matches_google
  GetGaData(myQuery)
}
}

