% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/deprecated.R
\name{GaFilter}
\alias{GaFilter}
\alias{GaFilter<-}
\title{GaFilter}
\usage{
GaFilter(...)

GaFilter(...) <- value
}
\arguments{
\item{...}{arguments passed onto \code{TableFilter}}

\item{value}{passed onto \code{TableFilter<-}}
}
\description{
Deprecated, use TableFilter instead.

Deprecated, use TableFilter<- instead.
}
\keyword{internal}

