% Generated by roxygen2 (4.0.1): do not edit by hand
\name{GaMaxResults}
\alias{GaMaxResults}
\alias{GaMaxResults<-}
\title{GaMaxResults}
\usage{
GaMaxResults(.Object)

GaMaxResults(.Object) <- value
}
\description{
Get the value set for MaxResults.

Set the maximum rows returned by a ganalytics query.
}

