% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/all-generics.R
\name{GaMaxResults}
\alias{GaMaxResults}
\alias{GaMaxResults<-}
\title{GaMaxResults}
\usage{
GaMaxResults(.Object)

GaMaxResults(.Object) <- value
}
\description{
Get the value set for MaxResults.

Set the maximum rows returned by a ganalytics query.
}

