% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/deprecated.R
\name{GaMaxResults}
\alias{GaMaxResults}
\alias{GaMaxResults<-}
\title{GaMaxResults}
\usage{
GaMaxResults(...)

GaMaxResults(...) <- value
}
\arguments{
\item{...}{arguments passed onto \code{MaxResults}}

\item{value}{passed onto \code{MaxResults<-}}
}
\description{
Deprecated, use MaxResults instead.

Deprecated, use MaxResults<- instead.
}
\keyword{internal}

