% Generated by roxygen2 (4.0.1): do not edit by hand
\name{GaMetaUpdate}
\alias{GaMetaUpdate}
\title{GaMetaUpdate
Update the metadata file
Update the package system file containing meta data about valid dimensions and metrics, etc.}
\usage{
GaMetaUpdate()
}
\value{
a data.frame
}
\description{
GaMetaUpdate
Update the metadata file
Update the package system file containing meta data about valid dimensions and metrics, etc.
}

