% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/meta.R
\name{GaMetaUpdate}
\alias{GaMetaUpdate}
\title{GaMetaUpdate
Update the metadata file
Update the package system file containing metadata about valid dimensions and metrics, etc.
This function should be used prior to package build.}
\usage{
GaMetaUpdate(creds = .creds)
}
\arguments{
\item{creds}{Google Analytics OAuth 2.0 credentials object.}
}
\value{
a data.frame
}
\description{
GaMetaUpdate
Update the metadata file
Update the package system file containing metadata about valid dimensions and metrics, etc.
This function should be used prior to package build.
}

