% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/all-generics.R
\name{GaMetrics}
\alias{GaMetrics}
\alias{GaMetrics<-}
\title{GaMetrics}
\usage{
GaMetrics(.Object, ...)

GaMetrics(.Object) <- value
}
\description{
Get the metrics of the object.

Set the metrics of the object.
}

