% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/deprecated.R
\name{GaMetrics}
\alias{GaMetrics}
\alias{GaMetrics<-}
\title{GaMetrics.}
\usage{
GaMetrics(...)

GaMetrics(...) <- value
}
\arguments{
\item{...}{arguments passed onto \code{Metrics}}

\item{value}{passed onto \code{Metrics}}
}
\description{
Deprecated, use \code{Metrics} instead.

Deprecated, use \code{Metrics<-} instead.
}
\keyword{internal}

