% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/var-classes.R
\docType{class}
\name{gaMetVar-class}
\alias{gaMetVar-class}
\title{`gaMetVar` class.}
\description{
An S4 class to represent a valid Google Analytics Core Reporting API metric
name.
}
\details{
Set to "ga:sessions" by default.
}
\keyword{internal}

