% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ganalytics-package.R
\docType{package}
\name{ganalytics}
\alias{ganalytics}
\alias{ganalytics-package}
\title{ganalytics}
\description{
A Google Analytics client for R
}
\details{
S4 classes and methods for interacting with Google Analytics data in R.
}
\references{
Google Analytics dimensions and metrics reference:
  \url{https://developers.google.com/analytics/devguides/reporting/core/dimsmets}
}
\keyword{package}

