\docType{data}
\name{ganalytics}
\alias{ganalytics}
\title{ganalytics}
\format{chr "https://accounts.google.com/o/oauth2/auth"}
\usage{
  auth_url
}
\description{
  Google Analytics client for R
}
\details{
  S4 classes and methods for interacting with Google
  Analytics with R.
}
\references{
  Google Analytics dimensions and metrics reference:
  \url{https://developers.google.com/analytics/devguides/reporting/core/dimsmets}
}
\keyword{datasets}

