% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/all-generics.R
\name{GaNonSequenceCondition}
\alias{GaNonSequenceCondition}
\title{GaNonSequenceCondition}
\usage{
GaNonSequenceCondition(.Object, ..., negation = FALSE)
}
\description{
Create a new gaNonSequenceCondition object
}

