% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GaSegment.R, R/all-generics.R
\docType{methods}
\name{GaNonSequenceCondition,.compoundExpr-method}
\alias{GaNonSequenceCondition}
\alias{GaNonSequenceCondition,.compoundExpr-method}
\title{GaNonSequenceCondition.}
\usage{
\S4method{GaNonSequenceCondition}{.compoundExpr}(object, ...,
  negation = FALSE)

GaNonSequenceCondition(object, ..., negation = FALSE)
}
\arguments{
\item{object}{An expression to be used as a non-sequential segment condition.}

\item{...}{Other expressions to be ANDed to the first expression provided.}

\item{negation}{Logical TRUE or FALSE to match segments where this conditon
  has not been met.}
}
\description{
Create a new gaConditionFilter object
}
\section{Methods (by class)}{
\itemize{
\item \code{.compoundExpr}: 
}}

