% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/deprecated.R
\name{GaProfileId}
\alias{GaProfileId}
\alias{GaProfileId<-}
\title{GaProfileId (Deprecated).}
\usage{
GaProfileId(...)

GaProfileId(...) <- value
}
\arguments{
\item{...}{arguments passed onto \code{GaView}}

\item{value}{passed onto \code{GaView<-}}
}
\description{
GaProfileId (Deprecated).

GaProfileId<- (Deprecated).
}
\keyword{internal}

