% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/deprecated.R
\name{GaProfileId}
\alias{GaProfileId}
\alias{GaProfileId<-}
\title{GaProfileId}
\usage{
GaProfileId(...)

GaProfileId(...) <- value
}
\arguments{
\item{...}{arguments passed onto \code{GaView}}

\item{value}{passed onto \code{GaView<-}}
}
\description{
Deprecated, use GaView instead.

Deprecated, use GaView instead.
}
\keyword{internal}

