% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/GaQuery.R
\name{GaQuery}
\alias{GaQuery}
\title{GaQuery
Create a ganalytics query object}
\usage{
GaQuery(view = GaAccounts()$entities[[1]], creds = list(),
  userName = character(0), appCreds = character(0),
  authFile = character(0), startDate = Sys.Date() - 8,
  endDate = Sys.Date() - 1, metrics = "ga:sessions",
  dimensions = "ga:date", sortBy = NULL, filters = NULL, segment = NULL,
  samplingLevel = "DEFAULT", maxResults = kGaMaxResults)
}
\arguments{
\item{authFile}{auth file to save token to}

\item{startDate}{start date}

\item{endDate}{end date}

\item{metrics}{character vector of metrics}

\item{dimensions}{character vector of dimensions}

\item{sortBy}{a sort by object}

\item{filters}{a filters object}

\item{segment}{a segment object}

\item{samplingLevel}{either "DEFAULT", "HIGHER_PRECISION" or "FASTER"}

\item{maxResults}{the maximum number of results to return,
up to 1,000,000}

\item{profileId}{profile id to use}
}
\description{
GaQuery
Create a ganalytics query object
}

