% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Query-methods.R
\name{GaQuery}
\alias{GaQuery}
\title{GaQuery}
\usage{
GaQuery(
  view = NA,
  creds = get_creds(),
  startDate = Sys.Date() - 8,
  endDate = Sys.Date() - 2,
  metrics = "ga:sessions",
  dimensions = "ga:date",
  sortBy = NULL,
  filters = NULL,
  segments = NULL,
  cohorts = NULL,
  samplingLevel = "DEFAULT",
  maxResults = kGaMaxResults
)
}
\arguments{
\item{view}{view id to use}

\item{creds}{authentication credentials object created using
\code{GoogleApiCreds()}}

\item{startDate}{start date}

\item{endDate}{end date}

\item{metrics}{character vector of metrics}

\item{dimensions}{character vector of dimensions}

\item{sortBy}{a sort by object}

\item{filters}{a filters object}

\item{segments}{a segment object or list of segments}

\item{cohorts}{a cohort object or a list of cohorts}

\item{samplingLevel}{either \code{"DEFAULT"}, \code{"HIGHER_PRECISION"} or
\code{"FASTER"}}

\item{maxResults}{the maximum number of results to return, up to 1,000,000}
}
\description{
Create a ganalytics query object
}
