% Generated by roxygen2 (4.0.2): do not edit by hand
\name{GaSamplingLevel}
\alias{GaSamplingLevel}
\alias{GaSamplingLevel<-}
\title{GaSamplingLevel}
\usage{
GaSamplingLevel(.Object)

GaSamplingLevel(.Object) <- value
}
\description{
Get the sampling level.

Set the sampling level for a ganalytics query.
}

