% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/deprecated.R
\name{GaSamplingLevel}
\alias{GaSamplingLevel}
\alias{GaSamplingLevel<-}
\title{GaSamplingLevel}
\usage{
GaSamplingLevel(...)

GaSamplingLevel(...) <- value
}
\arguments{
\item{...}{arguments passed onto \code{SamplingLevel}}

\item{value}{passed onto \code{SamplingLevel<-}}
}
\description{
Deprecated, use SamplingLevel instead.

Deprecated, use SamplingLevel<- instead.
}
\keyword{internal}

