% Generated by roxygen2 (4.0.2): do not edit by hand
\name{GaScopeLevel}
\alias{GaScopeLevel}
\alias{GaScopeLevel<-}
\title{GaScopeLevel}
\usage{
GaScopeLevel(.Object)

GaScopeLevel(.Object) <- value
}
\description{
Get the scope level of a gaDynSegment or gaMetExpr

Set the scope level of a gaDynSegment or a gaMetExpr
For segments, one of 'users' or 'sessions'
For metric expressions one of 'perUser', 'perSession' or 'perHit'
}

