% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/deprecated.R
\name{GaScopeLevel}
\alias{GaScopeLevel}
\alias{GaScopeLevel<-}
\title{GaScopeLevel}
\usage{
GaScopeLevel(...)

GaScopeLevel(...) <- value
}
\arguments{
\item{...}{passed onto \code{ScopeLevel}}

\item{value}{passed onto \code{ScopeLevel<-}}
}
\description{
Deprecated, use \code{ScopeLevel} instead

Deprecated, use \code{ScopeLevel<-} instead.
}
\keyword{internal}

