% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/all-generics.R
\name{GaSegment}
\alias{GaSegment}
\alias{GaSegment<-}
\title{GaSegment}
\usage{
GaSegment(.Object, ..., scope = "sessions")

GaSegment(.Object) <- value
}
\description{
Get the segment.

Set the segment
}

