% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{GaSegment}
\alias{GaSegment}
\alias{GaSegment<-}
\title{GaSegment}
\usage{
GaSegment(...)

GaSegment(...) <- value
}
\arguments{
\item{...}{arguments passed onto \code{Segments}}

\item{value}{passed onto \code{Segments}}
}
\description{
Deprecated, use Segments instead.

Deprecated, use Segments<- instead.
}
\keyword{internal}
