% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/deprecated.R
\name{GaSegment}
\alias{GaSegment}
\alias{GaSegment<-}
\title{GaSegment}
\usage{
GaSegment(...)

GaSegment(...) <- value
}
\arguments{
\item{...}{arguments passed onto \code{Segment}}

\item{value}{passed onto \code{Segment}}
}
\description{
Deprecated, use Segment instead.

Deprecated, use Segment<- instead.
}
\keyword{internal}

