% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GaSegment.R, R/all-generics.R
\docType{methods}
\name{GaSegmentCondition,.compoundExpr-method}
\alias{GaSegmentCondition}
\alias{GaSegmentCondition,.compoundExpr-method}
\alias{GaSegmentCondition,.gaSimpleOrSequence-method}
\alias{GaSegmentCondition,gaSegmentCondition-method}
\title{GaSegmentCondition.}
\usage{
\S4method{GaSegmentCondition}{.compoundExpr}(object, ..., scope = "sessions")

\S4method{GaSegmentCondition}{.gaSimpleOrSequence}(object, ...,
  scope = "sessions")

\S4method{GaSegmentCondition}{gaSegmentCondition}(object)

GaSegmentCondition(object, ..., scope = "sessions")
}
\arguments{
\item{object}{The first condition to be included in the segments definition.}

\item{...}{Other conditions to be included in the segment definition.}

\item{scope}{The scope of this condition, either 'user' or 'session' level.}
}
\description{
Create a new gaSegmentCondition object
}
\details{
A segment condition is either sequential or non-sequential. Sequential and
non-sequential conditoins can be combined using this function.
}
\section{Methods (by class)}{
\itemize{
\item \code{.compoundExpr}: 

\item \code{.gaSimpleOrSequence}: 

\item \code{gaSegmentCondition}: 
}}

