% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/all-generics.R
\name{GaSegmentCondition}
\alias{GaSegmentCondition}
\title{GaSegmentCondition}
\usage{
GaSegmentCondition(.Object, ..., scope = "sessions")
}
\description{
Create a new gaSegmentCondition object
}

