% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/segment-classes.R
\docType{class}
\name{gaSegmentConditionFilter-class}
\alias{gaSegmentConditionFilter-class}
\title{`gaSegmentConditionFilter` class.}
\description{
An S4 class to represent a non-sequential condition-based segment filter
expression.
}
\keyword{internal}

