% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/deprecated.R
\name{GaSegmentFilters}
\alias{GaSegmentFilters}
\title{GaSegmentFilters}
\usage{
GaSegmentFilters(..., scope = "sessions")
}
\arguments{
\item{...}{passed onto \code{PerSession} or \code{PerUser}}

\item{scope}{used to select either PerUser or PerSession.}
}
\description{
Deprecated, use \code{PerSession} or \code{PerUser} instead.
}
\keyword{internal}

