% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/segment-classes.R
\docType{class}
\name{gaSegmentSequenceFilter-class}
\alias{gaSegmentSequenceFilter-class}
\title{`gaSegmentSequenceFilter` class.}
\description{
An S4 class to represent a sequence expression as a segment filter.
}
\keyword{internal}

