% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/deprecated.R
\name{GaSequence}
\alias{GaSequence}
\title{GaSequence}
\usage{
GaSequence(...)
}
\arguments{
\item{...}{passed onto \code{Sequence}}
}
\description{
Deprecated, use \code{Sequence} instead.
}
\keyword{internal}

