% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/all-generics.R
\name{GaSequenceCondition}
\alias{GaSequenceCondition}
\title{GaSequenceCondition}
\usage{
GaSequenceCondition(object, ..., negation = FALSE)
}
\description{
Create a new gaSequenceCondition object
}

