% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/deprecated.R
\name{GaSortBy}
\alias{GaSortBy}
\alias{GaSortBy<-}
\title{GaSortBy (Deprecated).}
\usage{
GaSortBy(...)

GaSortBy(...) <- value
}
\arguments{
\item{...}{arguments passed onto \code{SortBy}}

\item{value}{passed onto \code{SortBy<-}}
}
\description{
GaSortBy (Deprecated).

GaSortBy<- (Deprecated).
}

