% Generated by roxygen2 (4.0.2): do not edit by hand
\name{GaSplitDateRange}
\alias{GaSplitDateRange}
\title{GaSplitDateRange
Splits a gaDateRange object into N pieces. Useful for splitting a query into
smaller chunks in order to overcome sampling.}
\usage{
GaSplitDateRange(dateRange, N)
}
\arguments{
\item{dateRange}{the gaDateRange object to be split}

\item{N}{the number of the separate date ranges to be split into; use 0 for
single days.}
}
\description{
GaSplitDateRange
Splits a gaDateRange object into N pieces. Useful for splitting a query into
smaller chunks in order to overcome sampling.
}

