% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/deprecated.R
\name{GaSplitDateRange}
\alias{GaSplitDateRange}
\title{GaSplitDateRange}
\usage{
GaSplitDateRange(...)
}
\arguments{
\item{...}{arguments passed onto \code{SplitDateRange}}
}
\description{
Deprecated, use SplitDateRange instead.
}
\keyword{internal}

