% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Date.R
\name{GaStartDate}
\alias{GaStartDate}
\alias{GaStartDate<-}
\title{GaStartDate (Deprecated)
For backwards compatibility}
\usage{
GaStartDate(...)

GaStartDate(...) <- value
}
\arguments{
\item{...}{arguments passed onto \code{StartDate}}

\item{value}{passed onto \code{StartDate}}
}
\description{
GaStartDate (Deprecated)
For backwards compatibility

GaStartDate<- (Deprecated)
}

