% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/deprecated.R
\name{GaStartDate}
\alias{GaStartDate}
\alias{GaStartDate<-}
\title{GaStartDate}
\usage{
GaStartDate(...)

GaStartDate(...) <- value
}
\arguments{
\item{...}{arguments passed onto \code{StartDate}}

\item{value}{passed onto \code{StartDate}}
}
\description{
Deprecated, use StartDate instead.

Deprecated, use StartDate<- instead.
}
\keyword{internal}

