% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/all-generics.R
\name{GaVar}
\alias{GaVar}
\alias{GaVar<-}
\alias{McfVar}
\title{GaVar}
\usage{
GaVar(.Object)

GaVar(.Object) <- value

McfVar(.Object)
}
\description{
Gets or Creates an object from the superclass .gaVar

Sets the value of an object or sets its slot belonging to the superclass .gaVar

Gets or Creates an object from the superclass .gaVar
}

