% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/all-generics.R
\name{GaVar}
\alias{GaVar}
\alias{GaVar<-}
\title{GaVar}
\usage{
GaVar(object)

GaVar(object) <- value
}
\description{
Gets or Creates an object from the superclass .gaVar

Sets the value of an object or sets its slot belonging to the superclass .gaVar
}

