% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/GetGa.R
\name{GetGaData}
\alias{GetGaData}
\title{GetGaData
Execute a ganalytics query.}
\usage{
GetGaData(object, creds,
  scope = "https://www.googleapis.com/auth/analytics.readonly")
}
\arguments{
\item{query}{the query to execute.}

\item{quiet}{supress messages}

\item{details}{show detailed messages}

\item{.progress}{progress bar to display. use .progress = "none" to turn off.}
}
\value{
a dataframe
}
\description{
GetGaData
Execute a ganalytics query.
}

