% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gtm-api-classes.R
\name{GtmAccount}
\alias{GtmAccount}
\title{GTM Account}
\usage{
GtmAccount(id = NULL, creds = .creds)
}
\arguments{
\item{id}{ID of the GTM account to get}

\item{creds}{The Google APIs credentials to use.}
}
\description{
Get a GTM account.
}

