% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gtm-api-classes.R
\name{GtmAccounts}
\alias{GtmAccounts}
\title{GTM Accounts}
\usage{
GtmAccounts(creds = .creds)
}
\arguments{
\item{creds}{The Google APIs credentials to use.}
}
\description{
Get a collection of GTM accounts.
}

