% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Segment-generics.R, R/Segment-methods.R
\docType{methods}
\name{Include}
\alias{Include}
\alias{Include,ANY-method}
\title{Include.}
\usage{
Include(object, scope = NULL)

\S4method{Include}{ANY}(object, scope = NULL)
}
\arguments{
\item{object}{a segment condition or sequence filter to include.}
}
\value{
a .gaSegmentFilter object with its negate slot set to FALSE.
}
\description{
Set the negation flag of a segment filter to FALSE.
}
\section{Methods (by class)}{
\itemize{
\item \code{ANY}: Define an include segment filter using the supplied
expression.
}}

