% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Segment-generics.R, R/Segment-methods.R
\docType{methods}
\name{Include}
\alias{Include}
\alias{Include,ANY-method}
\title{Include.}
\usage{
Include(object)

\S4method{Include}{ANY}(object)
}
\arguments{
\item{object}{a segment filter condition or sequence to include.}
}
\value{
a .gaSegmentFilter object with its negate slot set to FALSE.
}
\description{
Define a segment filter with its negation flag is set to FALSE.
}
\section{Methods (by class)}{
\itemize{
\item \code{ANY}: Define an include segment filter using the supplied
expression.
}}

