% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Segment-generics.R, R/Segment-methods.R
\docType{methods}
\name{Include}
\alias{Include}
\alias{Include,.compoundExpr-method}
\alias{Include,gaSegmentSequenceFilter-method}
\title{Include.}
\usage{
Include(object, ..., scope = "sessions")

\S4method{Include}{.compoundExpr}(object, ..., scope = "sessions")

\S4method{Include}{gaSegmentSequenceFilter}(object, ..., scope = "sessions")
}
\arguments{
\item{object}{a condition or sequence to include}

\item{...}{further conditions or sequences to include, i.e. ANDed.}

\item{scope}{the scope of the returned gaSegmentList, either "users" OR
"sessions".}
}
\value{
a .gaSegmentFilter object with its negate slot set to FALSE.
}
\description{
One or more segment conditions or sequences to include from the defined user
or session segment.
}
\section{Methods (by class)}{
\itemize{
\item \code{.compoundExpr}: Define an include segment filter using the supplied
expression.

\item \code{gaSegmentSequenceFilter}: Return the supplied Sequence Segment Filter with negation
not set so as to produce an include filter.
}}

