% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Segment.R, R/all-generics.R
\docType{methods}
\name{Include,.compoundExpr-method}
\alias{Include}
\alias{Include,.compoundExpr-method}
\alias{Include,gaSegmentSequenceFilter-method}
\title{Include.}
\usage{
\S4method{Include}{.compoundExpr}(object, ...)

\S4method{Include}{gaSegmentSequenceFilter}(object, ...)

Include(object, ...)
}
\arguments{
\item{object}{a condition or sequence to include}

\item{...}{further conditions or sequences to include, i.e. ANDed.}
}
\value{
a .gaSegmentFilter object with its negate slot set to FALSE.
}
\description{
One or more segment conditions or sequences to include from the defined user
or session segment.
}
\section{Methods (by class)}{
\itemize{
\item \code{.compoundExpr}: 

\item \code{gaSegmentSequenceFilter}: 
}}

