% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Segment-generics.R, R/Segment-methods.R
\docType{methods}
\name{IsNegated}
\alias{IsNegated}
\alias{IsNegated,.gaSegmentFilter-method}
\alias{IsNegated<-}
\alias{IsNegated<-,.gaSegmentFilter,logical-method}
\title{IsNegated.}
\usage{
IsNegated(object)

IsNegated(object) <- value

\S4method{IsNegated}{.gaSegmentFilter}(object)

\S4method{IsNegated}{.gaSegmentFilter,logical}(object) <- value
}
\arguments{
\item{object}{an object belonging to the superclass \code{.gaSegmentFilter}.}

\item{value}{the value of the negation slot, either \code{TRUE} or
  \code{FALSE}.}
}
\description{
Tests whether a segment filter is negated.

Sets whether a segment filter should be negated.
}
\section{Methods (by class)}{
\itemize{
\item \code{.gaSegmentFilter}: 

\item \code{object = .gaSegmentFilter,value = logical}: 
}}

