% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Segment-generics.R, R/Segment-methods.R
\name{IsNegated}
\alias{IsNegated}
\alias{IsNegated<-}
\alias{IsNegated,.gaSegmentFilter-method}
\alias{IsNegated<-,.gaSegmentFilter,logical-method}
\title{IsNegated}
\usage{
IsNegated(object)

IsNegated(object) <- value

\S4method{IsNegated}{.gaSegmentFilter}(object)

\S4method{IsNegated}{.gaSegmentFilter,logical}(object) <- value
}
\arguments{
\item{object}{An object belonging to the superclass \code{.gaSegmentFilter}.}

\item{value}{The value of the negation slot, either \code{TRUE} or
\code{FALSE}.}
}
\description{
Tests whether a segment filter is negated.

Sets the negation flag of a segment filter.
}
\section{Functions}{
\itemize{
\item \code{IsNegated,.gaSegmentFilter-method}: Tests whether a segment filter is negated, i.e. used to
define an exclude filter for the segment.

\item \code{IsNegated<-,.gaSegmentFilter,logical-method}: Sets whether a segment filter should be negated, i.e.
used as an exclude filter in a segment definition.
}}

\examples{
exclude_one_time_shoppers <- Exclude(
  Expr(~transactions == 1, metricScope = "perUser"),
  scope = "users"
)
IsNegated(exclude_one_time_shoppers) # TRUE

}
\seealso{
Other dynamic segment functions: 
\code{\link{DynSegment}()},
\code{\link{Exclude}()},
\code{\link{Include}()},
\code{\link{PerHit}()},
\code{\link{PerProduct}()},
\code{\link{PerSession}()},
\code{\link{PerUser}()},
\code{\link{ScopeLevel}()},
\code{\link{SegmentConditionFilter}()},
\code{\link{Segments}()},
\code{\link{Segment}()}
}
\concept{dynamic segment functions}
