% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Segment-generics.R, R/Segment-methods.R
\docType{methods}
\name{IsNegated}
\alias{IsNegated}
\alias{IsNegated<-}
\alias{IsNegated,.gaSegmentFilter-method}
\alias{IsNegated<-,.gaSegmentFilter,logical-method}
\title{IsNegated.}
\usage{
IsNegated(object)

IsNegated(object) <- value

\S4method{IsNegated}{.gaSegmentFilter}(object)

\S4method{IsNegated}{.gaSegmentFilter,logical}(object) <- value
}
\arguments{
\item{object}{an object belonging to the superclass \code{.gaSegmentFilter}.}

\item{value}{the value of the negation slot, either \code{TRUE} or
\code{FALSE}.}
}
\description{
Tests whether a segment filter is negated.

Sets whether a segment filter should be negated.
}
\section{Methods (by class)}{
\itemize{
\item \code{.gaSegmentFilter}: Test whether a segment filter is negated, i.e. used to
define an exclude filter for the segment.

\item \code{object = .gaSegmentFilter,value = logical}: Set whether a segment filter should be negated, i.e.
used as an exclude filter in a segment definition.
}}

