% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Segment-generics.R, R/Segment-methods.R
\docType{methods}
\name{IsNegated}
\alias{IsNegated}
\alias{IsNegated,.gaSegmentFilter-method}
\title{IsNegated}
\usage{
IsNegated(object)

\S4method{IsNegated}{.gaSegmentFilter}(object)
}
\arguments{
\item{object}{an object to test for negation belonging to the superclass
  \code{.gaSegmentFilter}.}
}
\description{
Tests whether a segment filter is negated.
}
\section{Methods (by class)}{
\itemize{
\item \code{.gaSegmentFilter}: 
}}

