% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Comparator-generics.R, R/Comparator-methods.R
\docType{methods}
\name{IsRegEx}
\alias{IsRegEx}
\alias{IsRegEx,.dimComparator-method}
\alias{IsRegEx,.expr-method}
\title{IsRegEx.}
\usage{
IsRegEx(object)

\S4method{IsRegEx}{.dimComparator}(object)

\S4method{IsRegEx}{.expr}(object)
}
\arguments{
\item{object}{An object to check if whether a regular expression.}
}
\value{
TRUE or FALSE
}
\description{
Checks for a regular expression.
}
\section{Methods (by class)}{
\itemize{
\item \code{.dimComparator}: 

\item \code{.expr}: 
}}

