% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Comparator-generics.R, R/Comparator-methods.R
\docType{methods}
\name{IsRegEx}
\alias{IsRegEx}
\alias{IsRegEx,.dimComparator-method}
\alias{IsRegEx,.expr-method}
\title{IsRegEx}
\usage{
IsRegEx(object)

\S4method{IsRegEx}{.dimComparator}(object)

\S4method{IsRegEx}{.expr}(object)
}
\arguments{
\item{object}{An object to check if whether a regular expression.}
}
\value{
\code{TRUE} or \code{FALSE}
}
\description{
Checks for a regular expression.
}
\section{Methods (by class)}{
\itemize{
\item \code{.dimComparator}: Test whether the supplied comparator is for a regular
expression.

\item \code{.expr}: Test whether a conditional expression is using regular
expression match.
}}

