% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{IsVarMatch}
\alias{IsVarMatch}
\title{IsVarMatch.}
\usage{
IsVarMatch(thisVar, inVars)
}
\arguments{
\item{thisVar}{var to compare against inVars.}

\item{inVars}{vector of vars to check var against.}
}
\description{
The following method is a temporary workaround to support XX placeholders in dimension and metric
names, such as with custom dimensions, metrics and various goal related variables.
}
\keyword{internal}

