% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/all-generics.R
\name{Later}
\alias{Later}
\title{Later.}
\usage{
Later(object, ...)
}
\arguments{
\item{object}{The expression that should preceed others in the sequence.}

\item{...}{Any other expressions that should follow the first one but before
  any others in the sequence.}
}
\value{
a gaSegmentSequenceStep object, with the immediate flag not set.
}
\description{
Treat a step within a sequence as happening at any point after any preceding
steps in the sequence, i.e. 'later'. 'Later' means 'followed by', but not
necessarily immediately.
}
\seealso{
Sequence
}

