% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/all-generics.R
\name{MaxResults}
\alias{MaxResults}
\alias{MaxResults<-}
\title{MaxResults}
\usage{
MaxResults(.Object)

MaxResults(.Object) <- value
}
\description{
Get the value set for MaxResults.

Set the maximum rows returned by a ganalytics query.
}

