% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Query-generics.R, R/Query-methods.R
\docType{methods}
\name{MaxResults}
\alias{MaxResults}
\alias{MaxResults<-}
\alias{MaxResults,.query,missing-method}
\alias{MaxResults,.query,ANY-method}
\alias{MaxResults<-,.query-method}
\title{MaxResults}
\usage{
MaxResults(object, value)

MaxResults(object) <- value

\S4method{MaxResults}{.query,missing}(object)

\S4method{MaxResults}{.query,ANY}(object, value)

\S4method{MaxResults}{.query}(object) <- value
}
\arguments{
\item{object}{A query object.}

\item{value}{Replacement value for the max-results parameter of the query.}
}
\description{
Get the value set for MaxResults.

Set the maximum rows returned by a ganalytics query.
}
\section{Methods (by class)}{
\itemize{
\item \code{object = .query,value = missing}: Return the maximum number of rows a query is allowed
to return.

\item \code{object = .query,value = ANY}: Set the maximum number of rows a query is allowed to
return.

\item \code{.query}: Set the maximum number of rows a query is allowed to
return.
}}

\seealso{
Other query object functions: \code{\link{Dimensions}},
  \code{\link{Metrics}}, \code{\link{SamplingLevel}},
  \code{\link{SortBy}}, \code{\link{TableFilter}}

Other query object functions: \code{\link{Dimensions}},
  \code{\link{Metrics}}, \code{\link{SamplingLevel}},
  \code{\link{SortBy}}, \code{\link{TableFilter}}
}
\concept{query object functions}
