% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Query.R, R/all-generics.R
\docType{methods}
\name{MaxResults,.query-method}
\alias{MaxResults}
\alias{MaxResults,.query-method}
\alias{MaxResults<-}
\alias{MaxResults<-,.query-method}
\title{MaxResults.}
\usage{
\S4method{MaxResults}{.query}(object)

\S4method{MaxResults}{.query}(object) <- value

MaxResults(object, value)

MaxResults(object) <- value
}
\arguments{
\item{object}{a query object.}

\item{value}{replacement value for the max-results parameter of the query.}
}
\description{
Get the value set for MaxResults.

Set the maximum rows returned by a ganalytics query.
}
\section{Methods (by class)}{
\itemize{
\item \code{.query}: 

\item \code{.query}: 
}}

