% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/comparator-classes.R
\docType{class}
\name{.mcfComparator-class}
\alias{.mcfComparator-class}
\title{`.mcfComparator` class.}
\description{
An S4 class union representing Multi-Channel Funnel condition comparator
operators.
}
\keyword{internal}

