% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/var-list-classes.R
\docType{class}
\name{mcfDimensions-class}
\alias{mcfDimensions-class}
\title{`mcfDimensions` class.}
\description{
An S4 class to represent a list of Multi-Channel Funnel dimensions.
}
\keyword{internal}

