% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/all-generics.R
\name{McfExpr}
\alias{McfExpr}
\title{McfExpr}
\usage{
McfExpr(.Object, operator, operand)
}
\description{
Create an expression.
}
\examples{
\donttest{
  myQuery <- McfQuery(view = 123456789)
  source_matches_google <- McfExpr("mcf:source", "~", "google")
  TableFilter(myQuery) <- source_matches_google
  GetData(myQuery)
}
}

