% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Expr.R, R/all-generics.R
\docType{methods}
\name{McfExpr,character,character-method}
\alias{McfExpr}
\alias{McfExpr,character,character-method}
\title{McfExpr.}
\usage{
\S4method{McfExpr}{character,character}(object, comparator, operand)

McfExpr(object, comparator, operand)
}
\arguments{
\item{object}{A dimension or metric variable, or another object to be coerced
to an .expr object.}

\item{comparator}{The comparator to use for the expression.}

\item{operand}{The operand to use for the expression.}
}
\description{
Create a Multi-Chanel Funnel Reporting API expression.
}
\section{Methods (by class)}{
\itemize{
\item \code{object = character,comparator = character}: 
}}
\examples{
\donttest{
  myQuery <- McfQuery(view = 123456789)
  source_matches_google <- McfExpr("mcf:source", "~", "google")
  TableFilter(myQuery) <- source_matches_google
  GetData(myQuery)
}
}
\seealso{
Other expression generators: \code{\link{Expr}},
  \code{\link{Expr,.expr,ANY-method}},
  \code{\link{Expr,character,character-method}},
  \code{\link{Expr,formula,ANY-method}};
  \code{\link{GaExpr}},
  \code{\link{GaExpr,character,character-method}};
  \code{\link{RtExpr}},
  \code{\link{RtExpr,character,character-method}}
}

