% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/operand-classes.R
\docType{class}
\name{mcfMetOperand-class}
\alias{mcfMetOperand-class}
\title{`mcfMetOperand` class.}
\description{
An S4 class to represent Multi-Channel metric condition operand.
}
\keyword{internal}

