% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/all-generics.R
\name{McfVar}
\alias{McfVar}
\alias{McfVar<-}
\title{McfVar}
\usage{
McfVar(object)

McfVar(object) <- value
}
\description{
Gets or Creates an object from the superclass .mcfVar

Sets the value of an object or sets its slot belonging to the superclass .mcfVar
}

