% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/comparator-classes.R
\docType{class}
\name{.metComparator-class}
\alias{.metComparator-class}
\title{`.metComparator` class.}
\description{
An S4 class union representing metric condition comparator operators.
}
\keyword{internal}

