% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/operand-classes.R
\docType{class}
\name{.metOperand-class}
\alias{.metOperand-class}
\title{`metOperand` class.}
\description{
An S4 class to represent the numerical value of a metric condition expression.
}
\keyword{internal}

