% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Var-list-generics.R, R/Var-methods.R
\docType{methods}
\name{Metrics}
\alias{Metrics}
\alias{Metrics<-}
\alias{Metrics,ANY-method}
\alias{Metrics,.query-method}
\alias{Metrics<-,.query-method}
\title{Metrics.}
\usage{
Metrics(object, ...)

Metrics(object) <- value

\S4method{Metrics}{ANY}(object, ...)

\S4method{Metrics}{.query}(object)

\S4method{Metrics}{.query}(object) <- value
}
\arguments{
\item{object}{An object to coerce to a list of metrics, or a query object to
replace the metrics of.}

\item{...}{Further metrics to add to the resulting list or the replacement
value for the metrics of the query object (if supplied).}

\item{value}{The replacement dimensions for the supplied object.}
}
\description{
Get the metrics of the object.

Set the metrics of the object.
}
\section{Methods (by class)}{
\itemize{
\item \code{ANY}: Coerce one or more supplied objects to .metrics.

\item \code{.query}: Get the list of metrics for a '.query'.

\item \code{.query}: Set the metrics for a '.query' object.
}}

