% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/all-generics.R
\name{Metrics}
\alias{Metrics}
\alias{Metrics<-}
\title{Metrics}
\usage{
Metrics(.Object, ...)

Metrics(.Object) <- value
}
\description{
Get the metrics of the object.

Set the metrics of the object.
}

