% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/all-generics.R
\name{Operand}
\alias{Operand}
\alias{Operand<-}
\title{Operand}
\usage{
Operand(object)

Operand(object) <- value
}
\description{
Get the operand of an expression.

Set the operand of an expression.
}

