% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Operand-generics.R, R/Operand-methods.R
\docType{methods}
\name{Operand}
\alias{Operand}
\alias{Operand,.expr-method}
\alias{Operand,.operand-method}
\alias{Operand,character-method}
\alias{Operand,logical-method}
\alias{Operand,numeric-method}
\alias{Operand<-}
\alias{Operand<-,.expr-method}
\title{Operand.}
\usage{
Operand(object, value)

Operand(object) <- value

\S4method{Operand}{.operand}(object)

\S4method{Operand}{.expr}(object)

\S4method{Operand}{character}(object)

\S4method{Operand}{numeric}(object)

\S4method{Operand}{logical}(object)

\S4method{Operand}{.expr}(object) <- value
}
\arguments{
\item{object}{The object for which to set the operand of.}

\item{value}{Character or numeric. The value to set the operand to.}
}
\description{
Get the operand of an expression.

Set the operand of an expression.
}
\section{Methods (by class)}{
\itemize{
\item \code{.operand}: Returns itself.

\item \code{.expr}: Return the operand used within the condition.

\item \code{character}: Coerce a character value as dimension operand.

\item \code{numeric}: Coerce a numeric value as a metric operand.

\item \code{logical}: Coerce a logical value as a dimension operator.

\item \code{.expr}: Replace the operand of a condition.
}}

