% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Operand-generics.R, R/Operand-methods.R
\docType{methods}
\name{Operand}
\alias{Operand}
\alias{Operand,.expr-method}
\alias{Operand,.operand-method}
\alias{Operand,character-method}
\alias{Operand,logical-method}
\alias{Operand,numeric-method}
\alias{Operand<-}
\alias{Operand<-,.expr-method}
\title{Operand.}
\usage{
Operand(object, value)

Operand(object) <- value

\S4method{Operand}{.operand}(object)

\S4method{Operand}{.expr}(object)

\S4method{Operand}{character}(object)

\S4method{Operand}{numeric}(object)

\S4method{Operand}{logical}(object)

\S4method{Operand}{.expr}(object) <- value
}
\arguments{
\item{object}{The object for which to set the operand of.}

\item{value}{Character or numeric. The value to set the operand to.}
}
\description{
Get the operand of an expression.

Set the operand of an expression.
}
\section{Methods (by class)}{
\itemize{
\item \code{.operand}: 

\item \code{.expr}: 

\item \code{character}: 

\item \code{numeric}: 

\item \code{logical}: 

\item \code{.expr}: 
}}

