% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Operand-generics.R, R/Operand-methods.R
\docType{methods}
\name{Operand}
\alias{Operand}
\alias{Operand<-}
\alias{Operand,.expr-method}
\alias{Operand<-,.expr-method}
\title{Operand.}
\usage{
Operand(object, value)

Operand(object) <- value

\S4method{Operand}{.expr}(object)

\S4method{Operand}{.expr}(object) <- value
}
\arguments{
\item{object}{The object for which to set the operand of.}

\item{value}{Character or numeric. The value to set the operand to.}
}
\description{
Get the operand of an expression.

Set the operand of an expression.
}
\section{Methods (by class)}{
\itemize{
\item \code{.expr}: Return the operand used within the condition, or coerce
the supplied value into an operand.

\item \code{.expr}: Replace the operand of a condition.
}}

