% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Operand.R, R/all-generics.R
\docType{methods}
\name{Operand,.operand-method}
\alias{Operand}
\alias{Operand,.expr-method}
\alias{Operand,.operand-method}
\alias{Operand<-}
\alias{Operand<-,.expr-method}
\title{Operand.}
\usage{
\S4method{Operand}{.operand}(object)

\S4method{Operand}{.expr}(object)

\S4method{Operand}{.expr}(object) <- value

Operand(object, value)

Operand(object) <- value
}
\arguments{
\item{object}{The object for which to set the operand of.}

\item{value}{Character or numeric. The value to set the operand to.}
}
\description{
Get the operand of an expression.

Set the operand of an expression.
}
\section{Methods (by class)}{
\itemize{
\item \code{.operand}: 

\item \code{.expr}: 

\item \code{.expr}: 
}}

