% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Operator.R, R/all-generics.R
\docType{methods}
\name{Operator,.operator-method}
\alias{Operator}
\alias{Operator,.expr-method}
\alias{Operator,.operator-method}
\alias{Operator<-}
\alias{Operator<-,.expr,ANY-method}
\alias{Operator<-,.operator,character-method}
\title{Operator.}
\usage{
\S4method{Operator}{.operator}(object)

\S4method{Operator}{.operator,character}(object) <- value

\S4method{Operator}{.expr}(object)

\S4method{Operator}{.expr,ANY}(object) <- value

Operator(object, ...)

Operator(object) <- value
}
\arguments{
\item{object}{The object to be coerced to a '.Operator' class or to get the
operator from.}

\item{value}{The value to set the operator to.}

\item{...}{Used by certain methods.}
}
\description{
Get or create an operator used in an expression.

Set the operator used in an expression.
}
\section{Methods (by class)}{
\itemize{
\item \code{.operator}: 

\item \code{object = .operator,value = character}: 

\item \code{.expr}: 

\item \code{object = .expr,value = ANY}: 
}}

