% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Bool.R, R/all-generics.R
\docType{methods}
\name{Or,.compoundExpr-method}
\alias{Or}
\alias{Or,.compoundExpr-method}
\alias{xor}
\alias{xor,.compoundExpr,.compoundExpr-method}
\alias{|,.compoundExpr,.compoundExpr-method}
\title{Or.}
\usage{
\S4method{Or}{.compoundExpr}(object, ...)

\S4method{|}{.compoundExpr,.compoundExpr}(e1, e2)

\S4method{xor}{.compoundExpr,.compoundExpr}(x, y)

Or(object, ...)

xor(x, y)
}
\arguments{
\item{object}{The first object to include within the OR expression.}

\item{...}{Additional objects to include within the OR expression.}

\item{e1}{first expression}

\item{e2}{second expression}

\item{x}{The first expression}

\item{y}{The second expression.}
}
\description{
OR two or more expressions.

Generate an expression that gives the exclusive or of two expressions.
}
\section{Methods (by class)}{
\itemize{
\item \code{.compoundExpr}: 

\item \code{e1 = .compoundExpr,e2 = .compoundExpr}: 

\item \code{x = .compoundExpr,y = .compoundExpr}: 
}}

