% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{parse_date}
\alias{parse_date}
\title{parse_date}
\usage{
parse_date(date, output_format = kGaDateInFormat)
}
\description{
Coerce a date into a character string formatted to either the
input or output format of the Google Analytics reporting API.
}
\keyword{internal}

