% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Segment-generics.R, R/Segment-methods.R
\docType{methods}
\name{PerHit}
\alias{PerHit}
\alias{PerHit,.compoundExpr-method}
\alias{PerHit,gaMetExpr-method}
\title{PerHit.}
\usage{
PerHit(object, ...)

\S4method{PerHit}{gaMetExpr}(object, ...)

\S4method{PerHit}{.compoundExpr}(object, ...)
}
\arguments{
\item{object}{a gaMetExpr object to coerce to user-level.}

\item{...}{Other conditions to be included in the segment definition.}
}
\description{
Set the scope of a gaMetExpr object to hit-level, or transforms a condition
filter to a sequence filter of length one (i.e. conditions to match a single
hit).
}
\section{Methods (by class)}{
\itemize{
\item \code{gaMetExpr}: Set the scope of the supplied metric condition to
hit-level.

\item \code{.compoundExpr}: Create a single step sequence filter from the supplied
expression.
}}

