% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Segment.R, R/all-generics.R
\docType{methods}
\name{PerSession,gaSegmentFilterList-method}
\alias{PerSession}
\alias{PerSession,gaMetExpr-method}
\alias{PerSession,gaSegmentFilterList-method}
\title{PerSession}
\usage{
\S4method{PerSession}{gaSegmentFilterList}(object, ...)

\S4method{PerSession}{gaMetExpr}(object, ...)

PerSession(object, ...)
}
\arguments{
\item{object}{a gaSegmentFilterList or gaMetExpr object to coerce to session-level.}

\item{...}{Other conditions to be included in the segment definition.}
}
\description{
Set the scope of a gaSegmentFilterList or gaMetExpr object to session-level.
}
\section{Methods (by class)}{
\itemize{
\item \code{gaSegmentFilterList}: 

\item \code{gaMetExpr}: 
}}

