% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Segment-generics.R, R/Segment-methods.R
\docType{methods}
\name{PerSession}
\alias{PerSession}
\alias{PerSession,ANY-method}
\alias{PerSession,gaMetExpr-method}
\alias{PerSession,gaSegmentFilterList-method}
\title{PerSession.}
\usage{
PerSession(object, ...)

\S4method{PerSession}{gaSegmentFilterList}(object, ...)

\S4method{PerSession}{ANY}(object, ...)

\S4method{PerSession}{gaMetExpr}(object, ...)
}
\arguments{
\item{object}{a gaSegmentFilterList or gaMetExpr object to coerce to session-level.}

\item{...}{Other conditions to be included in the segment definition.}
}
\description{
Set the scope of a gaSegmentFilterList or gaMetExpr object to session-level.
}
\section{Methods (by class)}{
\itemize{
\item \code{gaSegmentFilterList}: Set the scope of the supplied segment filters to
session level.

\item \code{ANY}: Create a session level segment filter list from the
supplied expressions, interpreted as condition filters.

\item \code{gaMetExpr}: Set the scope of the supplied metric condition to
session-level.
}}

