% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Segment-generics.R, R/Segment-methods.R
\docType{methods}
\name{PerUser}
\alias{PerUser}
\alias{PerUser,ANY-method}
\alias{PerUser,gaMetExpr-method}
\alias{PerUser,formula-method}
\title{PerUser.}
\usage{
PerUser(object, ...)

\S4method{PerUser}{ANY}(object, ...)

\S4method{PerUser}{gaMetExpr}(object, ...)

\S4method{PerUser}{formula}(object, ...)
}
\arguments{
\item{object}{a gaSegmentFilterList or gaMetExpr object to coerce to user-level.
Alternatively, an dimension expression or segment filter to coerce into a user
scoped gaSegmentFilterList.}

\item{...}{Other filters to include in the gaSegmentFilterList.}
}
\value{
a gaMetExpr or gaSegmentFilterList.

To define a gaSegmentFilterList using metric expressions rather than setting the
scope of the metric expression itself, wrap the metric expression in an
\code{Include} or \code{Exclude} call.
}
\description{
Set the scope of a gaSegmentFilterList or gaMetExpr object to user-level.
}
\section{Methods (by class)}{
\itemize{
\item \code{ANY}: Create a user-level segment filter list from the supplied
expressions, each interpreted as an include segment filter.

\item \code{gaMetExpr}: Set the scope of the supplied metric condition to
user-level.

\item \code{formula}: Set the scope of the supplied non-standard-evaluation
metric condition to user-level.
}}

