% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Segment-generics.R, R/Segment-methods.R
\docType{methods}
\name{PerUser}
\alias{PerUser}
\alias{PerUser,ANY-method}
\alias{PerUser,gaMetExpr-method}
\title{PerUser.}
\usage{
PerUser(object, ...)

\S4method{PerUser}{ANY}(object, ...)

\S4method{PerUser}{gaMetExpr}(object, ...)
}
\arguments{
\item{object}{a gaSegmentFilterList or gaMetExpr object to coerce to user-level.}

\item{...}{Other conditions to be included in the segment definition.}
}
\description{
Set the scope of a gaSegmentFilterList or gaMetExpr object to user-level.
}
\section{Methods (by class)}{
\itemize{
\item \code{ANY}: Create a user-level segment filter list from the supplied
expressions, each interpreted as an include segment filter.

\item \code{gaMetExpr}: Set the scope of the supplied metric condition to
user-level.
}}

