% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Segment.R, R/all-generics.R
\docType{methods}
\name{PerUser,gaSegmentFilterList-method}
\alias{PerUser}
\alias{PerUser,.compoundExpr-method}
\alias{PerUser,.gaSegmentFilter-method}
\alias{PerUser,gaMetExpr-method}
\alias{PerUser,gaSegmentFilterList-method}
\title{PerUser}
\usage{
\S4method{PerUser}{gaSegmentFilterList}(object, ...)

\S4method{PerUser}{.gaSegmentFilter}(object, ...)

\S4method{PerUser}{.compoundExpr}(object, ...)

\S4method{PerUser}{gaMetExpr}(object, ...)

PerUser(object, ...)
}
\arguments{
\item{object}{a gaSegmentFilterList or gaMetExpr object to coerce to user-level.}

\item{...}{Other conditions to be included in the segment definition.}
}
\description{
Set the scope of a gaSegmentFilterList or gaMetExpr object to user-level.
}
\section{Methods (by class)}{
\itemize{
\item \code{gaSegmentFilterList}: 

\item \code{.gaSegmentFilter}: 

\item \code{.compoundExpr}: 

\item \code{gaMetExpr}: 
}}

