% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/comparator-classes.R
\docType{class}
\name{.rtComparator-class}
\alias{.rtComparator-class}
\title{`.rtComparator` class.}
\description{
An S4 class union representing Real-Time condition comparator operators.
}
\keyword{internal}

