% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Expr.R, R/all-generics.R
\docType{methods}
\name{RtExpr,character,character-method}
\alias{RtExpr}
\alias{RtExpr,character,character-method}
\title{RtExpr.}
\usage{
\S4method{RtExpr}{character,character}(object, operator, operand)

RtExpr(object, operator, operand)
}
\arguments{
\item{object}{A dimension or metric variable, or another object to be coerced
to an .expr object.}

\item{operator}{The operator to use for the expression.}

\item{operand}{The operand to use for the expression.}
}
\description{
Create a Real-Time Reporting API expression.
}
\section{Methods (by class)}{
\itemize{
\item \code{object = character,operator = character}: 
}}
\examples{
\donttest{
  myQuery <- RtQuery(view = 123456789)
  source_matches_google <- RtExpr("rt:source", "~", "google")
  TableFilter(myQuery) <- source_matches_google
  GetData(myQuery)
}
}

