% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/all-generics.R
\name{RtExpr}
\alias{RtExpr}
\title{RtExpr}
\usage{
RtExpr(.Object, operator, operand)
}
\description{
Create an expression.
}
\examples{
\donttest{
  myQuery <- RtQuery(view = 123456789)
  source_matches_google <- RtExpr("rt:source", "~", "google")
  Filter(myQuery) <- source_matches_google
  GetData(myQuery)
}
}

