% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Expr.R, R/all-generics.R
\docType{methods}
\name{RtExpr,character,character-method}
\alias{RtExpr}
\alias{RtExpr,character,character-method}
\title{RtExpr.}
\usage{
\S4method{RtExpr}{character,character}(object, comparator, operand)

RtExpr(object, comparator, operand)
}
\arguments{
\item{object}{A dimension or metric variable, or another object to be coerced
to an .expr object.}

\item{comparator}{The comparator to use for the expression.}

\item{operand}{The operand to use for the expression.}
}
\description{
Create a Real-Time Reporting API expression.
}
\section{Methods (by class)}{
\itemize{
\item \code{object = character,comparator = character}: 
}}
\examples{
\donttest{
  myQuery <- RtQuery(view = 123456789)
  source_matches_google <- RtExpr("rt:source", "~", "google")
  TableFilter(myQuery) <- source_matches_google
  GetData(myQuery)
}
}
\seealso{
Other expression generators: \code{\link{Expr}},
  \code{\link{Expr,.expr,ANY-method}},
  \code{\link{Expr,character,character-method}},
  \code{\link{Expr,formula,ANY-method}};
  \code{\link{GaExpr}},
  \code{\link{GaExpr,character,character-method}};
  \code{\link{McfExpr}},
  \code{\link{McfExpr,character,character-method}}
}

