% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/var-classes.R
\docType{class}
\name{rtMetVar-class}
\alias{rtMetVar-class}
\title{`rtMetVar` class.}
\description{
An S4 class to represent a Real-Time reporting metric.
}
\details{
Set to "rt:pageviews" by default.
}
\keyword{internal}

