% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/all-generics.R
\name{SamplingLevel}
\alias{SamplingLevel}
\alias{SamplingLevel<-}
\title{SamplingLevel}
\usage{
SamplingLevel(object)

SamplingLevel(object) <- value
}
\description{
Get the sampling level.

Set the sampling level for a ganalytics query.
}

