% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Query-generics.R, R/Query.R
\docType{methods}
\name{SamplingLevel}
\alias{SamplingLevel}
\alias{SamplingLevel,.standardQuery-method}
\alias{SamplingLevel,data.frame-method}
\alias{SamplingLevel<-}
\alias{SamplingLevel<-,.standardQuery-method}
\title{SamplingLevel.}
\usage{
SamplingLevel(object, value)

SamplingLevel(object) <- value

\S4method{SamplingLevel}{.standardQuery}(object)

\S4method{SamplingLevel}{.standardQuery}(object) <- value

\S4method{SamplingLevel}{data.frame}(object)
}
\arguments{
\item{object}{the query or response to check the sampling level of.}

\item{value}{if \code{object} is a query, then use  value to set the sampling
  level to of that query.}
}
\description{
Get the sampling level.

Set the sampling level for a ganalytics query.
}
\section{Methods (by class)}{
\itemize{
\item \code{.standardQuery}: 

\item \code{.standardQuery}: 

\item \code{data.frame}: 
}}

