% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Query.R, R/all-generics.R, R/deprecated.R
\docType{methods}
\name{SamplingLevel,.standardQuery-method}
\alias{GaSamplingLevel}
\alias{GaSamplingLevel<-}
\alias{SamplingLevel}
\alias{SamplingLevel,.standardQuery-method}
\alias{SamplingLevel,data.frame-method}
\alias{SamplingLevel<-}
\alias{SamplingLevel<-,.standardQuery-method}
\title{SamplingLevel.}
\usage{
\S4method{SamplingLevel}{.standardQuery}(object)

\S4method{SamplingLevel}{.standardQuery}(object) <- value

\S4method{SamplingLevel}{data.frame}(object)

SamplingLevel(object, value)

SamplingLevel(object) <- value

GaSamplingLevel(...)

GaSamplingLevel(...) <- value
}
\arguments{
\item{object}{the query or response to check the sampling level of.}

\item{value}{if \code{object} is a query, then use  value to set the sampling
  level to of that query.}

\item{...}{arguments passed onto \code{SamplingLevel}}

\item{value}{passed onto \code{SamplingLevel<-}}
}
\description{
Get the sampling level.

Set the sampling level for a ganalytics query.

GaSamplingLevel (Deprecated).

GaSamplingLevel<- (Deprecated).
}
\section{Methods (by class)}{
\itemize{
\item \code{.standardQuery}: 

\item \code{.standardQuery}: 

\item \code{data.frame}: 
}}

