% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Expr.R, R/Segment.R, R/all-generics.R
\docType{methods}
\name{ScopeLevel,gaSegMetExpr-method}
\alias{ScopeLevel}
\alias{ScopeLevel,gaSegMetExpr-method}
\alias{ScopeLevel,gaSegmentFilterList-method}
\alias{ScopeLevel<-}
\alias{ScopeLevel<-,gaMetExpr,character-method}
\alias{ScopeLevel<-,gaSegMetExpr,character-method}
\alias{ScopeLevel<-,gaSegmentFilterList,character-method}
\title{ScopeLevel.}
\usage{
\S4method{ScopeLevel}{gaSegMetExpr}(object)

\S4method{ScopeLevel}{gaSegMetExpr,character}(object) <- value

\S4method{ScopeLevel}{gaMetExpr,character}(object) <- value

\S4method{ScopeLevel}{gaSegmentFilterList}(object)

\S4method{ScopeLevel}{gaSegmentFilterList,character}(object) <- value

ScopeLevel(object, value)

ScopeLevel(object) <- value
}
\arguments{
\item{object}{Segment condition or combined segment conditions or metric
expression.}

\item{value}{If a new scope level is supplied, then this function will return
  an updated copy of the supplied object with the new scope applied.}
}
\description{
Get the scope level of a gaDynSegment or gaMetExpr

Set the scope level of a gaDynSegment or a gaMetExpr
For segments, one of 'users' or 'sessions'
For metric expressions one of 'perUser', 'perSession' or 'perHit'
}
\section{Methods (by class)}{
\itemize{
\item \code{gaSegMetExpr}: 

\item \code{object = gaSegMetExpr,value = character}: 

\item \code{object = gaMetExpr,value = character}: 

\item \code{gaSegmentFilterList}: 

\item \code{object = gaSegmentFilterList,value = character}: 
}}

