% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Expr.R, R/GaSegment.R, R/all-generics.R
\docType{methods}
\name{GaScopeLevel,gaSegMetExpr-method}
\alias{GaNonSequenceCondition}
\alias{GaNonSequenceCondition,.compoundExpr-method}
\alias{GaScopeLevel}
\alias{GaScopeLevel,gaSegMetExpr-method}
\alias{GaScopeLevel,gaSegmentCondition-method}
\alias{GaScopeLevel<-}
\alias{GaScopeLevel<-,gaMetExpr,character-method}
\alias{GaScopeLevel<-,gaSegMetExpr,character-method}
\alias{GaScopeLevel<-,gaSegmentCondition,character-method}
\alias{GaSegmentCondition}
\alias{GaSegmentCondition,.compoundExpr-method}
\alias{GaSegmentCondition,.gaSimpleOrSequence-method}
\alias{GaSegmentCondition,gaSegmentCondition-method}
\alias{Segment}
\alias{Segment,.compoundExpr-method}
\alias{Segment,.gaSimpleOrSequence-method}
\alias{Segment,NULL-method}
\alias{Segment,character-method}
\alias{Segment,gaDynSegment-method}
\alias{Segment,gaFilter-method}
\alias{Segment,gaQuery-method}
\alias{Segment,gaSegmentCondition-method}
\alias{Segment,gaSegmentId-method}
\alias{Segment,gaUserSegment-method}
\alias{Segment,numeric-method}
\alias{Segment<-}
\alias{Segment<-,gaDynSegment,ANY-method}
\alias{Segment<-,gaDynSegment,andExpr-method}
\alias{Segment<-,gaQuery,ANY-method}
\alias{Segment<-,gaSegmentId,ANY-method}
\title{GaNonSequenceCondition.}
\usage{
\S4method{GaScopeLevel}{gaSegMetExpr}(object)

\S4method{GaScopeLevel}{gaSegMetExpr,character}(object) <- value

\S4method{GaScopeLevel}{gaMetExpr,character}(object) <- value

\S4method{GaScopeLevel}{gaSegmentCondition}(object)

\S4method{GaScopeLevel}{gaSegmentCondition,character}(object) <- value

\S4method{GaNonSequenceCondition}{.compoundExpr}(object, ...,
  negation = FALSE)

\S4method{GaSegmentCondition}{.compoundExpr}(object, ..., scope = "sessions")

\S4method{GaSegmentCondition}{.gaSimpleOrSequence}(object, ...,
  scope = "sessions")

\S4method{GaSegmentCondition}{gaSegmentCondition}(object)

\S4method{GaScopeLevel}{gaSegmentCondition}(object)

\S4method{GaScopeLevel}{gaSegmentCondition,character}(object) <- value

\S4method{Segment}{gaSegmentId}(object)

\S4method{Segment}{character}(object)

\S4method{Segment}{numeric}(object)

\S4method{Segment}{gaDynSegment}(object)

\S4method{Segment}{.compoundExpr}(object, ..., scope = "sessions")

\S4method{Segment}{gaSegmentCondition}(object, ..., scope = "sessions")

\S4method{Segment}{.gaSimpleOrSequence}(object, ..., scope = "sessions")

\S4method{Segment}{gaFilter}(object, ..., scope = "sessions")

\S4method{Segment}{`NULL`}(object)

\S4method{Segment}{gaDynSegment,andExpr}(object) <- value

\S4method{Segment}{gaDynSegment,ANY}(object) <- value

\S4method{Segment}{gaSegmentId,ANY}(object) <- value

\S4method{Segment}{gaQuery}(object)

\S4method{Segment}{gaQuery,ANY}(object) <- value

\S4method{Segment}{gaUserSegment}(object)

GaNonSequenceCondition(object, ..., negation = FALSE)

GaSegmentCondition(object, ..., scope = "sessions")

GaScopeLevel(object, ...)

GaScopeLevel(object) <- value

Segment(object, ..., scope = "sessions")

Segment(object) <- value
}
\description{
Create a new gaNonSequenceCondition object

Create a new gaSegmentCondition object

Get the scope level of a gaDynSegment or gaMetExpr

Set the scope level of a gaDynSegment or a gaMetExpr
For segments, one of 'users' or 'sessions'
For metric expressions one of 'perUser', 'perSession' or 'perHit'

Get the segment.

Set the segment
}
\section{Methods (by class)}{
\itemize{
\item \code{gaSegMetExpr}: 

\item \code{object = gaSegMetExpr,value = character}: 

\item \code{object = gaMetExpr,value = character}: 

\item \code{gaSegmentCondition}: 

\item \code{object = gaSegmentCondition,value = character}: 

\item \code{.compoundExpr}: 

\item \code{.compoundExpr}: 

\item \code{.gaSimpleOrSequence}: 

\item \code{gaSegmentCondition}: 

\item \code{gaSegmentCondition}: 

\item \code{object = gaSegmentCondition,value = character}: 

\item \code{gaSegmentId}: 

\item \code{character}: 

\item \code{numeric}: 

\item \code{gaDynSegment}: 

\item \code{.compoundExpr}: 

\item \code{gaSegmentCondition}: 

\item \code{.gaSimpleOrSequence}: 

\item \code{gaFilter}: Coerce a Table Filter into a Segment

\item \code{NULL}: 

\item \code{object = gaDynSegment,value = andExpr}: 

\item \code{object = gaDynSegment,value = ANY}: 

\item \code{object = gaSegmentId,value = ANY}: 

\item \code{gaQuery}: 

\item \code{object = gaQuery,value = ANY}: 

\item \code{gaUserSegment}: 
}}

