% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Segment-generics.R, R/deprecated.R
\docType{data}
\name{Segment}
\alias{Segment}
\alias{Segment}
\alias{Segment<-}
\title{Segment.}
\format{An object of class \code{NULL} of length 0.}
\usage{
Segment(...)

Segment(...) <- value
}
\arguments{
\item{...}{other segment conditions, filters or filter lists to include
in the segment's definition (ANDed)}

\item{value}{passed onto \code{Segments}}

\item{object}{a segment or other object that can be coerced into
a segment, including dynamic segments, built-in and/or custom
segments by their ID.}

\item{...}{arguments passed onto \code{Segments}}
}
\value{
an object belonging to the .gaSegment superclass.
}
\description{
Define a segment for use in a query's segment list.

Deprecated, use Segments instead.

Deprecated, use Segments<- instead.
}
\keyword{datasets}
\keyword{internal}
