% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Segment-generics.R
\name{Segment}
\alias{Segment}
\title{Segment.}
\usage{
Segment(object, ...)
}
\arguments{
\item{object}{a segment or other object that can be coerced into
a segment, including dynamic segments, built-in and/or custom
segments by their ID.}

\item{...}{other segment conditions, filters or filter lists to include
in the segment's definition (ANDed)}
}
\value{
an object belonging to the .gaSegment superclass.
}
\description{
Define a segment for use in a query's segment list.
}
