% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Segment.R, R/all-generics.R
\docType{methods}
\name{Segment,gaSegmentId-method}
\alias{Segment}
\alias{Segment,.compoundExpr-method}
\alias{Segment,.gaSegmentFilter-method}
\alias{Segment,NULL-method}
\alias{Segment,character-method}
\alias{Segment,gaDynSegment-method}
\alias{Segment,gaFilter-method}
\alias{Segment,gaQuery-method}
\alias{Segment,gaSegmentFilterList-method}
\alias{Segment,gaSegmentId-method}
\alias{Segment,gaUserSegment-method}
\alias{Segment,numeric-method}
\alias{Segment<-}
\alias{Segment<-,gaDynSegment,ANY-method}
\alias{Segment<-,gaDynSegment,andExpr-method}
\alias{Segment<-,gaQuery,ANY-method}
\alias{Segment<-,gaSegmentId,ANY-method}
\title{Segment.}
\usage{
\S4method{Segment}{gaSegmentId}(object)

\S4method{Segment}{character}(object)

\S4method{Segment}{numeric}(object)

\S4method{Segment}{gaDynSegment}(object)

\S4method{Segment}{.compoundExpr}(object, ..., scope = "sessions")

\S4method{Segment}{gaSegmentFilterList}(object, ..., scope = "sessions")

\S4method{Segment}{.gaSegmentFilter}(object, ..., scope = "sessions")

\S4method{Segment}{gaFilter}(object, ..., scope = "sessions")

\S4method{Segment}{`NULL`}(object)

\S4method{Segment}{gaDynSegment,andExpr}(object) <- value

\S4method{Segment}{gaDynSegment,ANY}(object) <- value

\S4method{Segment}{gaSegmentId,ANY}(object) <- value

\S4method{Segment}{gaQuery}(object)

\S4method{Segment}{gaQuery,ANY}(object) <- value

\S4method{Segment}{gaUserSegment}(object)

Segment(object, ..., scope = "sessions")

Segment(object) <- value
}
\arguments{
\item{object}{An expression to coerce to a segment definition or segment ID}

\item{...}{Other expressions to combine with the first expression, if
appropriate.}

\item{scope}{The scope level to apply to the resulting segment definition.}

\item{value}{The segment definition or ID to set the segment parameter to.}
}
\description{
Get the segment.

Set the segment
}
\section{Methods (by class)}{
\itemize{
\item \code{gaSegmentId}: 

\item \code{character}: 

\item \code{numeric}: 

\item \code{gaDynSegment}: 

\item \code{.compoundExpr}: 

\item \code{gaSegmentFilterList}: 

\item \code{.gaSegmentFilter}: 

\item \code{gaFilter}: Coerce a Table Filter into a Segment

\item \code{NULL}: 

\item \code{object = gaDynSegment,value = andExpr}: 

\item \code{object = gaDynSegment,value = ANY}: 

\item \code{object = gaSegmentId,value = ANY}: 

\item \code{gaQuery}: 

\item \code{object = gaQuery,value = ANY}: 

\item \code{gaUserSegment}: 
}}

