% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Segment-generics.R, R/Segment-methods.R
\docType{methods}
\name{SegmentConditionFilter}
\alias{SegmentConditionFilter}
\alias{SegmentConditionFilter,.compoundExpr-method}
\alias{SegmentConditionFilter,gaSegmentConditionFilter-method}
\title{SegmentConditionFilter.}
\usage{
SegmentConditionFilter(object, ..., negation = FALSE)

\S4method{SegmentConditionFilter}{.compoundExpr}(object, ...,
  negation = FALSE)

\S4method{SegmentConditionFilter}{gaSegmentConditionFilter}(object)
}
\arguments{
\item{object}{An expression to be used as a non-sequential segment condition.}

\item{...}{Other expressions to be ANDed to the first expression provided.}

\item{negation}{Logical TRUE or FALSE to match segments where this conditon
has not been met.}
}
\value{
a gaSegmentConditionFilter object.
}
\description{
Create a new gaSegmentConditionFilter object
}
\section{Methods (by class)}{
\itemize{
\item \code{.compoundExpr}: Create a non-sequential segment condition
filter from one or more expressions. All conditions within the filter must
hold true within a single session.

\item \code{gaSegmentConditionFilter}: Returns itself.
}}

