% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Segment-generics.R, R/Segment-methods.R
\docType{methods}
\name{SegmentConditionFilter}
\alias{SegmentConditionFilter}
\alias{SegmentConditionFilter,ANY-method}
\title{SegmentConditionFilter.}
\usage{
SegmentConditionFilter(object, ..., negation = FALSE, scope = "sessions")

\S4method{SegmentConditionFilter}{ANY}(object, ..., negation = FALSE,
  scope = "sessions")
}
\arguments{
\item{object}{An expression to be used as a non-sequential segment condition.}

\item{...}{Other expressions to be ANDed to the first expression provided.}

\item{negation}{Logical TRUE or FALSE to match segments where this conditon
has not been met.}
}
\value{
a gaSegmentConditionFilter object.
}
\description{
Create a new gaSegmentConditionFilter object
}
\section{Methods (by class)}{
\itemize{
\item \code{ANY}: Create a non-sequential segment condition
filter from one or more expressions. All conditions within the filter must
hold true within a single session if applied to a gaSegmentFilterList
scoped at session-level, or to a single hit if scoped at user-level.
}}

