% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Segment.R, R/all-generics.R
\docType{methods}
\name{SegmentConditionFilter,.compoundExpr-method}
\alias{SegmentConditionFilter}
\alias{SegmentConditionFilter,.compoundExpr-method}
\title{SegmentConditionFilter.}
\usage{
\S4method{SegmentConditionFilter}{.compoundExpr}(object, ...,
  negation = FALSE)

SegmentConditionFilter(object, ..., negation = FALSE)
}
\arguments{
\item{object}{An expression to be used as a non-sequential segment condition.}

\item{...}{Other expressions to be ANDed to the first expression provided.}

\item{negation}{Logical TRUE or FALSE to match segments where this conditon
has not been met.}
}
\value{
a gaSegmentConditionFilter object.
}
\description{
Create a new gaSegmentConditionFilter object
}
\section{Methods (by class)}{
\itemize{
\item \code{.compoundExpr}: 
}}

