% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{SegmentFilters}
\alias{SegmentFilters}
\title{SegmentFilters}
\usage{
SegmentFilters(..., scope = "sessions")
}
\arguments{
\item{...}{passed onto \code{PerSession} or \code{PerUser}}

\item{scope}{used to select either PerUser or PerSession.}
}
\description{
Deprecated, use \code{PerSession} or \code{PerUser} instead.
}
\keyword{internal}
