% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Segment-generics.R, R/Segment.R
\docType{methods}
\name{SegmentFilters}
\alias{SegmentFilters}
\alias{SegmentFilters,.compoundExpr-method}
\alias{SegmentFilters,.gaSegmentFilter-method}
\alias{SegmentFilters,gaSegmentFilterList-method}
\title{SegmentFilters.}
\usage{
SegmentFilters(object, ..., scope = "sessions")

\S4method{SegmentFilters}{.compoundExpr}(object, ..., scope = "sessions")

\S4method{SegmentFilters}{.gaSegmentFilter}(object, ..., scope = "sessions")

\S4method{SegmentFilters}{gaSegmentFilterList}(object)
}
\arguments{
\item{object}{The first condition to be included in the segments definition.}

\item{...}{Other conditions to be included in the segment definition.}

\item{scope}{The scope of this condition, either 'user' or 'session' level.}
}
\value{
a gaSegmentFilterList object.
}
\description{
Create a new gaSegmentFilterList object
}
\details{
A segment condition is either sequential or non-sequential. Sequential and
non-sequential conditoins can be combined using this function.
}
\section{Methods (by class)}{
\itemize{
\item \code{.compoundExpr}: 

\item \code{.gaSegmentFilter}: 

\item \code{gaSegmentFilterList}: 
}}

