% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Segment-generics.R, R/Segment-methods.R
\docType{methods}
\name{Segments}
\alias{Segments}
\alias{Segments<-}
\alias{Segment,numeric-method}
\alias{Segment,character-method}
\alias{Segment,ANY-method}
\alias{Segment,NULL-method}
\alias{Segment,gaUserSegment-method}
\alias{Segments,gaSegmentList-method}
\alias{Segments,gaQuery-method}
\alias{Segments<-,gaQuery-method}
\title{Segments.}
\usage{
Segments(object, ...)

Segments(object) <- value

\S4method{Segment}{numeric}(object)

\S4method{Segment}{character}(object)

\S4method{Segment}{ANY}(object, ...)

\S4method{Segment}{`NULL`}(object)

\S4method{Segment}{gaUserSegment}(object)

\S4method{Segments}{gaSegmentList}(object)

\S4method{Segments}{gaQuery}(object)

\S4method{Segments}{gaQuery}(object) <- value
}
\arguments{
\item{object}{A query object to extract the segment list from.}

\item{...}{Alternatively, provide one or more named arguments
(segments or objects that can be coerced into segments)
including dynamic segments, built-in and/or custom segments by
their ID.}

\item{value}{The segment definition or ID or a list of segments.}
}
\value{
a gaSegmentList
}
\description{
Get the list of segments from the object or coerce the supplied objects into a
a named list of segments.

Set the segments of the query object.
}
\section{Methods (by class)}{
\itemize{
\item \code{numeric}: Interpret the supplied numeric value as a segment ID.

\item \code{character}: Interpret the supplied character value as a segment ID.

\item \code{ANY}: Create a non-sequential segment using the supplied
expressions.

\item \code{NULL}: returns NULL

\item \code{gaUserSegment}: Return the segment ID of the supplied GA Management API
user segment.

\item \code{gaSegmentList}: Returns itself

\item \code{gaQuery}: Return the definition of the segment applied to the view.

\item \code{gaQuery}: Set the segments to be used witin a query.
}}

