% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Segment.R, R/all-generics.R
\docType{methods}
\name{Later,.compoundExpr-method}
\alias{First,.compoundExpr-method}
\alias{Later,.compoundExpr-method}
\alias{Sequence}
\alias{Sequence,.compoundExpr-method}
\alias{Then,.compoundExpr-method}
\title{Sequence.}
\usage{
\S4method{Later}{.compoundExpr}(object)

\S4method{Then}{.compoundExpr}(object)

\S4method{First}{.compoundExpr}(object)

\S4method{Sequence}{.compoundExpr}(object, ..., negation = FALSE)

Sequence(object, ..., negation = FALSE)
}
\arguments{
\item{object}{A sequence step or another expression that should be coerced to
a sequence condition.}

\item{...}{Other steps within the sequence condition, in the order in which
they should be applied.}

\item{negation}{Logical TRUE or FALSE to match segments where this sequence
  has not occured.}
}
\description{
Create a new gaSequence object
}
\section{Methods (by class)}{
\itemize{
\item \code{.compoundExpr}: 

\item \code{.compoundExpr}: 

\item \code{.compoundExpr}: 

\item \code{.compoundExpr}: 
}}

