% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence.R
\name{sequential_segment}
\alias{sequential_segment}
\title{sequential_segment.}
\usage{
sequential_segment(steps)
}
\arguments{
\item{steps}{a list of expressions (of one or more conditions), each
repeating a step in the sequence.}
}
\description{
Create a sequence using R non-standard evaluation syntax.
}
\details{
Steps must be separated by commas (\code{,}). \code{...} denotes zero or more
interactions may precede the step that follows, otherwise without \code{...} then
the there must not be any interactions between the adjacent steps in order for the
results to the sequence.
}
\examples{
a <- Expr(~pagePath == "/home")
b <- Expr(~eventCategory == "Video") &
  Expr(~eventAction == "Play")
c <- Expr(~medium == "email")
s <- sequential_segment(list( ..., a, ..., b, c ))

}
