% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sequence.R
\name{sequential_segment}
\alias{sequential_segment}
\title{sequential_segment.}
\usage{
sequential_segment(steps)
}
\arguments{
\item{steps}{a list of expressions (of one or more conditions), each
  repreating a step in the sequence.}
}
\description{
Create a sequence using R non-standard evaluation syntax.
}
\details{
Steps must be separated by commas (\code{,}). ... denotes zero or more of
interactions may preceed the step that follows, otherwise without ... then
the there must not be any interactions between the adjacent steps.
}
\examples{
a <- Expr(~pagePath == "/home")
b <- Expr(~eventCategory == "Video") &
  Expr(~eventAction == "Play")
c <- Expr(~medium == "email")
s <- sequential_segment(list( ..., a, ..., b, c ))
}

