% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SortBy.R, R/all-generics.R
\docType{methods}
\name{SortBy,.varList-method}
\alias{GaSortBy}
\alias{GaSortBy<-}
\alias{SortBy}
\alias{SortBy,.query-method}
\alias{SortBy,.varList-method}
\alias{SortBy,NULL-method}
\alias{SortBy,character-method}
\alias{SortBy,list-method}
\alias{SortBy<-}
\alias{SortBy<-,.query-method}
\title{SortBy.}
\usage{
\S4method{SortBy}{.varList}(object, desc = logical(0))

\S4method{SortBy}{`NULL`}(object)

\S4method{SortBy}{character}(object, ..., desc = logical(0))

\S4method{SortBy}{list}(object, ..., desc = logical(0))

\S4method{SortBy}{.query}(object, value)

\S4method{SortBy}{.query}(object) <- value

SortBy(object, ..., desc = logical(0))

SortBy(object) <- value

SortBy(object, ..., desc = logical(0))

SortBy(object) <- value
}
\description{
Get the sortBy order of the query.

Set the order of rows returned by Google Analytics.
}
\section{Methods (by class)}{
\itemize{
\item \code{.varList}: 

\item \code{NULL}: 

\item \code{character}: 

\item \code{list}: 

\item \code{.query}: 

\item \code{.query}: 
}}

