% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SortBy.R, R/all-generics.R
\docType{methods}
\name{SortBy,.varList-method}
\alias{SortBy}
\alias{SortBy,.query-method}
\alias{SortBy,.varList-method}
\alias{SortBy,NULL-method}
\alias{SortBy,character-method}
\alias{SortBy,list-method}
\alias{SortBy<-}
\alias{SortBy<-,.query-method}
\title{SortBy.}
\usage{
\S4method{SortBy}{.varList}(object, desc = logical(0))

\S4method{SortBy}{`NULL`}(object)

\S4method{SortBy}{character}(object, ..., desc = logical(0))

\S4method{SortBy}{list}(object, ..., desc = logical(0))

\S4method{SortBy}{.query}(object, value)

\S4method{SortBy}{.query}(object) <- value

SortBy(object, ..., desc = logical(0))

SortBy(object) <- value
}
\arguments{
\item{object}{A character vector or list of dimensions or metrics to sort by.
If character, then prefixing the dimension name with a "+" means ascending
order or "-" for decending order. By default metrics are sorted in
descending order, while dimensions are by default in ascending order.
Alternatively, supply a query object and replacement dimensions and metrics
or sort by.}

\item{desc}{A logical vector, same length as the resulting list of dimension
  or metric variables, indicating which columns of the resulting query
  response should be sorted in decending order.}

\item{...}{further dimensions or metrics to sort by, or if \code{object} is a
query then the replacement list of dimensions or metrics to sort by.}

\item{value}{The replacement dimensions and metrics for the supplied object.}
}
\description{
Get the sortBy order of the query.

Set the order of rows returned by Google Analytics.
}
\section{Methods (by class)}{
\itemize{
\item \code{.varList}: 

\item \code{NULL}: 

\item \code{character}: 

\item \code{list}: 

\item \code{.query}: 

\item \code{.query}: 
}}

