% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Date-methods.R
\name{SplitDateRange}
\alias{SplitDateRange}
\title{SplitDateRange}
\usage{
SplitDateRange(dateRange, N = 0L)
}
\arguments{
\item{dateRange}{the \code{gaDateRange} object to be split}

\item{N}{the number of the separate date ranges to be split into; use
\code{0} for single days.}
}
\description{
Splits a gaDateRange object into \code{N} pieces. Useful for splitting a
query into smaller chunks in order to overcome sampling.
}
\seealso{
Other date range functions: \code{\link{Cohort}},
  \code{\link{DateRange}}, \code{\link{EndDate}},
  \code{\link{StartDate}}
}
\concept{date range functions}
