% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/query-classes.R
\docType{class}
\name{.standardQuery-class}
\alias{.standardQuery-class}
\title{`.standardQuery` class.}
\description{
An S4 class to represent a standard reporting API query.
}
\keyword{internal}

