% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Date-generics.R
\name{StartDate}
\alias{StartDate}
\alias{StartDate<-}
\title{StartDate}
\usage{
StartDate(object, value)

StartDate(object) <- value
}
\arguments{
\item{object}{Object to get start date of.}

\item{value}{Value to set start date of object to.}
}
\description{
Get the start date.

Set the start date.
}
\seealso{
Other date range functions: \code{\link{Cohort}},
  \code{\link{DateRange}}, \code{\link{EndDate}}

Other date range functions: \code{\link{Cohort}},
  \code{\link{DateRange}}, \code{\link{EndDate}}
}
\concept{date range functions}
