% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TableFilter-generics.R, R/TableFilter-methods.R
\docType{methods}
\name{TableFilter}
\alias{TableFilter}
\alias{TableFilter,.compoundExpr-method}
\alias{TableFilter,.query-method}
\alias{TableFilter,.tableFilter-method}
\alias{TableFilter,NULL-method}
\alias{TableFilter,gaDynSegment-method}
\alias{TableFilter<-}
\alias{TableFilter<-,.tableFilter,andExpr-method}
\alias{TableFilter<-,.query-method}
\title{TableFilter.}
\usage{
TableFilter(object, value)

TableFilter(object) <- value

\S4method{TableFilter}{.query}(object)

\S4method{TableFilter}{`NULL`}(object)

\S4method{TableFilter}{.tableFilter}(object)

\S4method{TableFilter}{.compoundExpr}(object)

\S4method{TableFilter}{gaDynSegment}(object)

\S4method{TableFilter}{.tableFilter,andExpr}(object) <- value
\S4method{TableFilter}{.query}(object) <- value
}
\arguments{
\item{object}{The object to be coerced to a TableFilter or the query object
to apply a table filter to.}

\item{value}{The replacement table filter where \code{object} is a query.}
}
\description{
Get the filter.

Set the filter.
}
\section{Methods (by class)}{
\itemize{
\item \code{.query}: Return the TableFilter that has been applied to the
given query

\item \code{NULL}: Returns NULL

\item \code{.tableFilter}: Returns itself

\item \code{.compoundExpr}: Coerce an expression into a table filter.

\item \code{gaDynSegment}: Coerce a segment into a table filter. Only possible
for segments without sequential conditions.

\item \code{.query}: Method to replace the table filter of a query
}}

