% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Filter.R, R/all-generics.R
\docType{methods}
\name{TableFilter,.query-method}
\alias{GaFilter}
\alias{GaFilter<-}
\alias{TableFilter}
\alias{TableFilter,.compoundExpr-method}
\alias{TableFilter,.query-method}
\alias{TableFilter,.tableFilter-method}
\alias{TableFilter,NULL-method}
\alias{TableFilter,gaDynSegment-method}
\alias{TableFilter<-}
\alias{TableFilter<-,.query,ANY-method}
\alias{TableFilter<-,.tableFilter,andExpr-method}
\title{TableFilter.}
\usage{
\S4method{TableFilter}{.query}(object)

\S4method{TableFilter}{`NULL`}(object)

\S4method{TableFilter}{.tableFilter}(object)

\S4method{TableFilter}{.compoundExpr}(object)

\S4method{TableFilter}{gaDynSegment}(object)

\S4method{TableFilter}{.tableFilter,andExpr}(object) <- value

\S4method{TableFilter}{.query,ANY}(object) <- value

TableFilter(object, ...)

TableFilter(object) <- value

TableFilter(object, ...)

TableFilter(object) <- value
}
\description{
Get the filter.

Set the filter.
}
\section{Methods (by class)}{
\itemize{
\item \code{.query}: 

\item \code{NULL}: 

\item \code{.tableFilter}: 

\item \code{.compoundExpr}: 

\item \code{gaDynSegment}: 

\item \code{object = .tableFilter,value = andExpr}: 

\item \code{object = .query,value = ANY}: 
}}

