% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/all-generics.R
\name{Then}
\alias{Then}
\title{Then.}
\usage{
Then(object, ...)
}
\arguments{
\item{object}{The expression that should \bold{immediately} preceed others in
the sequence.}

\item{...}{Any other expressions that should \bold{immediately} follow the
  first one but before any others in the sequence.}
}
\value{
a gaSegmentSequenceStep object, with the immediate flag set.
}
\description{
Treat a step within a sequence as happening immediately after any preceding
steps in the sequence, i.e. 'immediately following'.
}
\seealso{
Sequence
}

