% Generated by roxygen2 (4.0.1): do not edit by hand
\name{ValidGaOperand}
\alias{ValidGaOperand}
\title{ValidGaOperand}
\usage{
ValidGaOperand(gaVar, gaOperand)
}
\arguments{
\item{gaVar}{selected dimension to check operand against}

\item{gaOperand}{the operand value to check}
}
\description{
Checks whether an operand value is valid for a selected dimension.
}

