% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{ValidGaOperand}
\alias{ValidGaOperand}
\title{ValidGaOperand.}
\usage{
ValidGaOperand(var, operand)
}
\arguments{
\item{var}{selected dimension to check operand against}

\item{operand}{the operand value to check}
}
\description{
Checks whether an operand value is valid for a selected dimension.
}

