% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Bool.R, R/all-generics.R
\docType{methods}
\name{xor,.compoundExpr,.compoundExpr-method}
\alias{xor}
\alias{xor,.compoundExpr,.compoundExpr-method}
\title{Generate an expression that gives the Exclusive-OR of two expressions.}
\usage{
\S4method{xor}{.compoundExpr,.compoundExpr}(x, y)

xor(x, y)
}
\arguments{
\item{x,y}{Conditions for an exclusive-or expression.}
}
\description{
Generate an expression that gives the Exclusive-OR of two expressions.
}
\section{Methods (by class)}{
\itemize{
\item \code{x = .compoundExpr,y = .compoundExpr}: 
}}
\seealso{
\code{link{Or}} \code{link{And}} \code{link{Not}}
}

